Compute a Bayesian version of R-squared for regression models

Compute a Bayesian version of R-squared for regression models

bayes_R2.stanreg(object, newdata = NULL, re.form = NULL, offset = NULL,
  ...)

Arguments

object
A fitted model object returned by one of the rstanarm modeling functions. See stanreg-objects.

newdata
Similar to the newdata argument to posterior_linpred and posterior_predict except, in addition to new observations of the predictors, new observations of the outcome must be also included. See the Examples section below.

re.form, offset
For models with group-level terms, these arguments are passed to posterior_linpred if the newdata argument is specified.

...
Currently ignored.

Value

A vector of Bayesian R-squared values with length equal to the posterior sample size.

Examples

fit <- stan_glm(mpg ~ wt + cyl, data = mtcars)

#> trying deprecated constructor; please alert package maintainer

#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 1).
#> 
#> Gradient evaluation took 2e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.2 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:    1 / 2000 [  0%]  (Warmup)
#> Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Iteration: 2000 / 2000 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.098417 seconds (Warm-up)
#>                0.07896 seconds (Sampling)
#>                0.177377 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 2).
#> 
#> Gradient evaluation took 2e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.2 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:    1 / 2000 [  0%]  (Warmup)
#> Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Iteration: 2000 / 2000 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.092174 seconds (Warm-up)
#>                0.082216 seconds (Sampling)
#>                0.17439 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 3).
#> 
#> Gradient evaluation took 2.1e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.21 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:    1 / 2000 [  0%]  (Warmup)
#> Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Iteration: 2000 / 2000 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.091541 seconds (Warm-up)
#>                0.102294 seconds (Sampling)
#>                0.193835 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 4).
#> 
#> Gradient evaluation took 1.6e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.16 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:    1 / 2000 [  0%]  (Warmup)
#> Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Iteration: 2000 / 2000 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.106522 seconds (Warm-up)
#>                0.083153 seconds (Sampling)
#>                0.189675 seconds (Total)
#> 

rsq <- bayes_R2(fit)

#> Error in bayes_R2(fit): could not find function "bayes_R2"

print(median(rsq))

#> Error in median(rsq): object 'rsq' not found

# specifying newdata (including outcome variable 'mpg')
nd <- data.frame(mpg = c(10, 20, 30), wt = c(4, 3, 2), cyl = c(8, 6, 4))
rsq_new <- bayes_R2(fit, newdata = nd)

#> Error in bayes_R2(fit, newdata = nd): could not find function "bayes_R2"

print(median(rsq_new))

#> Error in median(rsq_new): object 'rsq_new' not found

# multilevel binomial model
if (!exists("example_model")) example(example_model)

#> Warning: no help found for ‘example_model’

print(example_model)

#> Error in print(example_model): object 'example_model' not found

median(bayes_R2(example_model))

#> Error in bayes_R2(example_model): could not find function "bayes_R2"

median(bayes_R2(example_model, re.form = NA)) # exclude group-level

#> Error in bayes_R2(example_model, re.form = NA): could not find function "bayes_R2"