get_y {rstanarm}	R Documentation

Extract X, Y or Z from a stanreg object

Description

Extract X, Y or Z from a stanreg object

Usage

get_y(object, ...)

get_x(object, ...)

get_z(object, ...)

Arguments

object
A fitted model object returned by one of the rstanarm modeling functions. See stanreg-objects.

...
Other arguments passed to methods. For a stanmvreg object this can be an integer m specifying the submodel.

Value

For get_x and get_z, a matrix. For get_y, either a vector or a matrix, depending on how the response variable was specified.

[Package rstanarm version 2.17.3 Index]