Bayesian ARIMA modeling of time series

Bayesian ARIMA modeling of time series

stan_arima(x, order = c(0L, 0L, 0L), seasonal = list(order = c(0L, 0L, 0L),
  period = NA), xreg = NULL, include.mean = TRUE, ..., prior = normal(),
  prior_intercept = normal(), algorithm = c("sampling", "optimizing",
  "meanfield", "fullrank"), adapt_delta = NULL)

Arguments

x, order, seasonal, xreg, include.mean
Same as arima.

...
Further arguments passed to the function in the rstan package (sampling, vb, or optimizing), corresponding to the estimation method named by algorithm. For example, if algorithm is "sampling" it is possibly to specify iter, chains, cores, refresh, etc.

algorithm
A string (possibly abbreviated) indicating the estimation approach to use. Can be "sampling" for MCMC (the default), "optimizing" for optimization, "meanfield" for variational inference with independent normal distributions, or "fullrank" for variational inference with a multivariate normal distribution. See rstanarm-package for more details on the estimation algorithms. NOTE: not all fitting functions support all four algorithms.

adapt_delta
Only relevant if algorithm="sampling". See adapt_delta for details.

Value

A stanreg object is returned for stan_arima.

A stanfit object (or a slightly modified stanfit object) is returned if stan_arima.fit is called directly.

Details

The stan_arima function is similar in syntax to arima but rather than performing maximum likelihood estimation, full Bayesian estimation is performed (if algorithm is "sampling") via MCMC.

See also

stanreg-methods and arima.

The vignette for stan_arima.

Examples

# add example of stan_arima