Extract the posterior sample

For models fit using MCMC (algorithm="sampling"), the posterior sample ---the post-warmup draws from the posterior distribution--- can be extracted from a fitted model object as a matrix, data frame, or array. The as.matrix and as.data.frame methods merge all chains together, whereas the as.array method keeps the chains separate. For models fit using optimization ("optimizing") or variational inference ("meanfield" or "fullrank"), there is no posterior sample but rather a matrix (or data frame) of 1000 draws from either the asymptotic multivariate Gaussian sampling distribution of the parameters or the variational approximation to the posterior distribution.

# S3 method for stanreg
as.matrix(x, ..., pars = NULL, regex_pars = NULL)

# S3 method for stanreg
as.array(x, ..., pars = NULL, regex_pars = NULL)

# S3 method for stanreg
as.data.frame(x, ..., pars = NULL, regex_pars = NULL)

Arguments

x
A fitted model object returned by one of the rstanarm modeling functions. See stanreg-objects.

...
Ignored.

pars
An optional character vector of parameter names.

regex_pars
An optional character vector of regular expressions to use for parameter selection. regex_pars can be used in place of pars or in addition to pars. Currently, all functions that accept a regex_pars argument ignore it for models fit using optimization.

Value

A matrix, data.frame, or array, the dimensions of which depend on pars and regex_pars, as well as the model and estimation algorithm (see the Description section above).

See also

Examples

# \donttest{
if (!exists("example_model")) example(example_model)

#> 
#> exmpl_> example_model <- 
#> exmpl_+   stan_glmer(cbind(incidence, size - incidence) ~ size + period + (1|herd),
#> exmpl_+              data = lme4::cbpp, family = binomial, QR = TRUE,
#> exmpl_+              # this next line is only to keep the example small in size!
#> exmpl_+              chains = 2, cores = 1, seed = 12345, iter = 500, refresh = 0)

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> 
#> exmpl_> example_model
#> stan_glmer
#>  family:       binomial [logit]
#>  formula:      cbind(incidence, size - incidence) ~ size + period + (1 | herd)
#>  observations: 56
#> ------
#>             Median MAD_SD
#> (Intercept) -1.5    0.6  
#> size         0.0    0.0  
#> period2     -1.0    0.3  
#> period3     -1.1    0.3  
#> period4     -1.6    0.4  
#> 
#> Error terms:
#>  Groups Name        Std.Dev.
#>  herd   (Intercept) 0.76    
#> Num. levels: herd 15 
#> 
#> ------
#> * For help interpreting the printed output see ?print.stanreg
#> * For info on the priors used see ?prior_summary.stanreg

# Extract posterior sample after MCMC
draws <- as.matrix(example_model)
print(dim(draws))

#> [1] 500  21

# For example, we can see that the median of the draws for the intercept 
# is the same as the point estimate rstanarm uses
print(median(draws[, "(Intercept)"]))

#> [1] -1.538909

print(example_model$coefficients[["(Intercept)"]])

#> [1] -1.538909

# The as.array method keeps the chains separate
draws_array <- as.array(example_model)
print(dim(draws_array)) # iterations x chains x parameters

#> [1] 250   2  21

# Extract draws from asymptotic Gaussian sampling distribution 
# after optimization
fit <- stan_glm(mpg ~ wt, data = mtcars, algorithm = "optimizing")
draws <- as.data.frame(fit)
print(colnames(draws))

#> [1] "(Intercept)" "wt"          "sigma"      

print(nrow(draws)) # 1000 draws are taken

#> [1] 1000

# Extract draws from variational approximation to the posterior distribution
fit2 <- update(fit, algorithm = "meanfield")

#> Chain 1: ------------------------------------------------------------
#> Chain 1: EXPERIMENTAL ALGORITHM:
#> Chain 1:   This procedure has not been thoroughly tested and may be unstable
#> Chain 1:   or buggy. The interface is subject to change.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Gradient evaluation took 1.8e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.18 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Begin eta adaptation.
#> Chain 1: Iteration:   1 / 250 [  0%]  (Adaptation)
#> Chain 1: Iteration:  50 / 250 [ 20%]  (Adaptation)
#> Chain 1: Iteration: 100 / 250 [ 40%]  (Adaptation)
#> Chain 1: Iteration: 150 / 250 [ 60%]  (Adaptation)
#> Chain 1: Iteration: 200 / 250 [ 80%]  (Adaptation)
#> Chain 1: Success! Found best value [eta = 1] earlier than expected.
#> Chain 1: 
#> Chain 1: Begin stochastic gradient ascent.
#> Chain 1:   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes 
#> Chain 1:    100         -129.863             1.000            1.000
#> Chain 1:    200         -107.240             0.605            1.000
#> Chain 1:    300          -90.886             0.464            0.211
#> Chain 1:    400          -90.923             0.348            0.211
#> Chain 1:    500          -92.955             0.283            0.180
#> Chain 1:    600          -90.491             0.240            0.180
#> Chain 1:    700          -90.626             0.206            0.027
#> Chain 1:    800          -92.299             0.183            0.027
#> Chain 1:    900          -90.679             0.164            0.022
#> Chain 1:   1000          -90.786             0.148            0.022
#> Chain 1:   1100          -90.800             0.048            0.018
#> Chain 1:   1200          -90.845             0.027            0.018
#> Chain 1:   1300          -90.568             0.009            0.003
#> Chain 1:   1400          -90.578             0.009            0.003
#> Chain 1:   1500          -90.730             0.007            0.002
#> Chain 1:   1600          -90.573             0.005            0.002
#> Chain 1:   1700          -91.414             0.005            0.002
#> Chain 1:   1800          -90.588             0.004            0.002
#> Chain 1:   1900          -90.620             0.003            0.002
#> Chain 1:   2000          -90.781             0.003            0.002
#> Chain 1:   2100          -90.577             0.003            0.002
#> Chain 1:   2200          -90.497             0.003            0.002
#> Chain 1:   2300          -90.989             0.003            0.002
#> Chain 1:   2400          -91.774             0.004            0.002
#> Chain 1:   2500          -90.849             0.005            0.005
#> Chain 1:   2600          -91.042             0.005            0.005
#> Chain 1:   2700          -90.585             0.005            0.005
#> Chain 1:   2800          -90.714             0.004            0.002
#> Chain 1:   2900          -90.667             0.004            0.002
#> Chain 1:   3000          -90.715             0.004            0.002
#> Chain 1:   3100          -90.466             0.004            0.003
#> Chain 1:   3200          -90.807             0.004            0.004
#> Chain 1:   3300          -90.923             0.004            0.003
#> Chain 1:   3400          -90.695             0.003            0.003
#> Chain 1:   3500          -90.617             0.002            0.002
#> Chain 1:   3600          -90.743             0.002            0.001
#> Chain 1:   3700          -90.902             0.002            0.001
#> Chain 1:   3800          -90.645             0.002            0.002
#> Chain 1:   3900          -91.108             0.002            0.003
#> Chain 1:   4000          -91.436             0.003            0.003
#> Chain 1:   4100          -90.666             0.003            0.003
#> Chain 1:   4200          -90.831             0.003            0.003
#> Chain 1:   4300          -91.838             0.004            0.003
#> Chain 1:   4400          -90.453             0.005            0.004
#> Chain 1:   4500          -90.770             0.005            0.004
#> Chain 1:   4600          -90.625             0.005            0.004
#> Chain 1:   4700          -90.567             0.005            0.004
#> Chain 1:   4800          -90.757             0.005            0.004
#> Chain 1:   4900          -90.597             0.005            0.003
#> Chain 1:   5000          -90.527             0.005            0.002
#> Chain 1:   5100          -90.533             0.004            0.002
#> Chain 1:   5200          -90.547             0.004            0.002
#> Chain 1:   5300          -90.481             0.003            0.002
#> Chain 1:   5400          -90.610             0.001            0.001
#> Chain 1:   5500          -90.558             0.001            0.001
#> Chain 1:   5600          -90.575             0.001            0.001
#> Chain 1:   5700          -91.058             0.001            0.001
#> Chain 1:   5800          -91.149             0.001            0.001
#> Chain 1:   5900          -90.672             0.002            0.001
#> Chain 1:   6000          -90.767             0.002            0.001
#> Chain 1:   6100          -90.774             0.002            0.001
#> Chain 1:   6200          -90.666             0.002            0.001
#> Chain 1:   6300          -90.544             0.002            0.001
#> Chain 1:   6400          -90.444             0.002            0.001
#> Chain 1:   6500          -90.665             0.002            0.001
#> Chain 1:   6600          -90.587             0.002            0.001
#> Chain 1:   6700          -90.633             0.001            0.001
#> Chain 1:   6800          -90.655             0.001            0.001
#> Chain 1:   6900          -90.475             0.001            0.001
#> Chain 1:   7000          -90.658             0.001            0.001
#> Chain 1:   7100          -90.531             0.001            0.001
#> Chain 1:   7200          -91.076             0.002            0.001
#> Chain 1:   7300          -90.483             0.002            0.002
#> Chain 1:   7400          -90.615             0.002            0.002
#> Chain 1:   7500          -90.960             0.002            0.002
#> Chain 1:   7600          -90.683             0.003            0.002
#> Chain 1:   7700          -90.655             0.003            0.002
#> Chain 1:   7800          -91.085             0.003            0.003
#> Chain 1:   7900          -91.095             0.003            0.003
#> Chain 1:   8000          -90.377             0.004            0.004
#> Chain 1:   8100          -90.866             0.004            0.005
#> Chain 1:   8200          -90.729             0.003            0.004
#> Chain 1:   8300          -90.687             0.003            0.003
#> Chain 1:   8400          -90.713             0.003            0.003
#> Chain 1:   8500          -90.390             0.003            0.003
#> Chain 1:   8600          -90.634             0.003            0.003
#> Chain 1:   8700          -90.518             0.003            0.003
#> Chain 1:   8800          -90.715             0.003            0.002
#> Chain 1:   8900          -90.740             0.003            0.002
#> Chain 1:   9000          -90.646             0.002            0.002
#> Chain 1:   9100          -90.695             0.001            0.001
#> Chain 1:   9200          -90.646             0.001            0.001
#> Chain 1:   9300          -90.561             0.001            0.001
#> Chain 1:   9400          -90.641             0.001            0.001
#> Chain 1:   9500          -90.670             0.001            0.001
#> Chain 1:   9600          -90.482             0.001            0.001
#> Chain 1:   9700          -90.650             0.001            0.001
#> Chain 1:   9800          -90.455             0.001            0.001
#> Chain 1:   9900          -91.058             0.002            0.001
#> Chain 1:   10000          -90.506             0.002            0.002
#> Chain 1: Informational Message: The maximum number of iterations is reached! The algorithm may not have converged.
#> Chain 1: This variational approximation is not guaranteed to be meaningful.
#> Chain 1: 
#> Chain 1: Drawing a sample of size 1000 from the approximate posterior... 
#> Chain 1: COMPLETED.

#> Warning: Pareto k diagnostic value is 1.15. Resampling is disabled. Decreasing tol_rel_obj may help if variational algorithm has terminated prematurely. Otherwise consider using sampling instead.

#> Setting 'QR' to TRUE can often be helpful when using one of the variational inference algorithms. See the documentation for the 'QR' argument.

draws <- as.data.frame(fit2, pars = "wt")
print(colnames(draws))

#> [1] "wt"

print(nrow(draws)) # 1000 draws are taken

#> [1] 1000

# }