Extract the posterior sample

For models fit using MCMC (algorithm="sampling"), the posterior sample ---the post-warmup draws from the posterior distribution--- can be extracted from a fitted model object as a matrix, data frame, or array. The as.matrix and as.data.frame methods merge all chains together, whereas the as.array method keeps the chains separate. For models fit using optimization ("optimizing") or variational inference ("meanfield" or "fullrank"), there is no posterior sample but rather a matrix (or data frame) of 1000 draws from either the asymptotic multivariate Gaussian sampling distribution of the parameters or the variational approximation to the posterior distribution.

# S3 method for stanreg
as.matrix(x, ..., pars = NULL, regex_pars = NULL)

# S3 method for stanreg
as.array(x, ..., pars = NULL, regex_pars = NULL)

# S3 method for stanreg
as.data.frame(x, ..., pars = NULL, regex_pars = NULL)

Arguments

x
A fitted model object returned by one of the rstanarm modeling functions. See stanreg-objects.

...
Ignored.

pars
An optional character vector of parameter names.

regex_pars
An optional character vector of regular expressions to use for parameter selection. regex_pars can be used in place of pars or in addition to pars. Currently, all functions that accept a regex_pars argument ignore it for models fit using optimization.

Value

A matrix, data.frame, or array, the dimensions of which depend on pars and regex_pars, as well as the model and estimation algorithm (see the Description section above).

See also

stanreg-methods

Examples

if (!exists("example_model")) example(example_model)

#> Warning: no help found for ‘example_model’

# Extract posterior sample after MCMC
draws <- as.matrix(example_model)

#> Error in as.matrix(example_model): object 'example_model' not found

print(dim(draws))

#> Error in print(dim(draws)): object 'draws' not found

# For example, we can see that the median of the draws for the intercept 
# is the same as the point estimate rstanarm uses
print(median(draws[, "(Intercept)"]))

#> Error in median(draws[, "(Intercept)"]): object 'draws' not found

print(example_model$coefficients[["(Intercept)"]])

#> Error in print(example_model$coefficients[["(Intercept)"]]): object 'example_model' not found

# The as.array method keeps the chains separate
draws_array <- as.array(example_model)

#> Error in as.array(example_model): object 'example_model' not found

print(dim(draws_array)) # iterations x chains x parameters

#> Error in print(dim(draws_array)): object 'draws_array' not found

# Extract draws from asymptotic Gaussian sampling distribution 
# after optimization
fit <- stan_glm(mpg ~ wt, data = mtcars, algorithm = "optimizing")

#> trying deprecated constructor; please alert package maintainer

#> Initial log joint probability = -198.855
#> Optimization terminated normally: 
#>   Convergence detected: relative gradient magnitude is below tolerance

draws <- as.data.frame(fit)
print(colnames(draws))

#> [1] "(Intercept)" "wt"          "sigma"      

print(nrow(draws)) # 1000 draws are taken

#> [1] 1000

# Extract draws from variational approximation to the posterior distribution
fit2 <- update(fit, algorithm = "meanfield")

#> trying deprecated constructor; please alert package maintainer

#> ------------------------------------------------------------
#> EXPERIMENTAL ALGORITHM:
#>   This procedure has not been thoroughly tested and may be unstable
#>   or buggy. The interface is subject to change.
#> ------------------------------------------------------------
#> 
#> 
#> 
#> Gradient evaluation took 2.1e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.21 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Begin eta adaptation.
#> Iteration:   1 / 250 [  0%]  (Adaptation)
#> Iteration:  50 / 250 [ 20%]  (Adaptation)
#> Iteration: 100 / 250 [ 40%]  (Adaptation)
#> Iteration: 150 / 250 [ 60%]  (Adaptation)
#> Iteration: 200 / 250 [ 80%]  (Adaptation)
#> Success! Found best value [eta = 1] earlier than expected.
#> 
#> Begin stochastic gradient ascent.
#>   iter       ELBO   delta_ELBO_mean   delta_ELBO_med   notes 
#>    100     -1e+02             1.000            1.000
#>    200     -1e+02             0.542            1.000
#>    300     -1e+02             0.387            0.084
#>    400     -1e+02             0.313            0.093
#>    500     -1e+02             0.288            0.093
#>    600     -9e+01             0.249            0.093
#>    700     -9e+01             0.214            0.084
#>    800     -9e+01             0.188            0.084
#>    900     -9e+01             0.168            0.076
#>   1000     -9e+01             0.153            0.076
#>   1100     -9e+01             0.053            0.055
#>   1200     -9e+01             0.045            0.016
#>   1300     -9e+01             0.039            0.013
#>   1400     -9e+01             0.031            0.012
#>   1500     -9e+01             0.012            0.011
#>   1600     -9e+01             0.007            0.006   MEAN ELBO CONVERGED   MEDIAN ELBO CONVERGED
#> 
#> Drawing a sample of size 1000 from the approximate posterior... 
#> COMPLETED.

#> Setting 'QR' to TRUE can often be helpful when using one of the variational inference algorithms. See the documentation for the 'QR' argument.

draws <- as.data.frame(fit2, pars = "wt")
print(colnames(draws))

#> [1] "wt"

print(nrow(draws)) # 1000 draws are taken

#> [1] 1000