Example model

A model for use in rstanarm examples.

Format

Calling example("example_model") will run the model in the Examples section, below, and the resulting stanreg object will then be available in the global environment. The chains and iter arguments are specified to make this example be small in size. In practice, we recommend that they be left unspecified in order to use the default values (4 and 2000 respectively) or increased if there are convergence problems. The cores argument is optional and on a multicore system, the user may well want to set that equal to the number of chains being executed.

See also

cbpp for a description of the data.

Examples

example_model <-
  stan_glmer(cbind(incidence, size - incidence) ~ size + period + (1|herd),
             data = lme4::cbpp, family = binomial,
             # this next line is only to keep the example small in size!
             chains = 2, cores = 1, seed = 12345, iter = 500)

#> trying deprecated constructor; please alert package maintainer

#> 
#> SAMPLING FOR MODEL 'binomial' NOW (CHAIN 1).
#> 
#> Gradient evaluation took 7.2e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.72 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:   1 / 500 [  0%]  (Warmup)
#> Iteration:  50 / 500 [ 10%]  (Warmup)
#> Iteration: 100 / 500 [ 20%]  (Warmup)
#> Iteration: 150 / 500 [ 30%]  (Warmup)
#> Iteration: 200 / 500 [ 40%]  (Warmup)
#> Iteration: 250 / 500 [ 50%]  (Warmup)
#> Iteration: 251 / 500 [ 50%]  (Sampling)
#> Iteration: 300 / 500 [ 60%]  (Sampling)
#> Iteration: 350 / 500 [ 70%]  (Sampling)
#> Iteration: 400 / 500 [ 80%]  (Sampling)
#> Iteration: 450 / 500 [ 90%]  (Sampling)
#> Iteration: 500 / 500 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.448672 seconds (Warm-up)
#>                0.15146 seconds (Sampling)
#>                0.600132 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'binomial' NOW (CHAIN 2).
#> 
#> Gradient evaluation took 9.7e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.97 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:   1 / 500 [  0%]  (Warmup)
#> Iteration:  50 / 500 [ 10%]  (Warmup)
#> Iteration: 100 / 500 [ 20%]  (Warmup)
#> Iteration: 150 / 500 [ 30%]  (Warmup)
#> Iteration: 200 / 500 [ 40%]  (Warmup)
#> Iteration: 250 / 500 [ 50%]  (Warmup)
#> Iteration: 251 / 500 [ 50%]  (Sampling)
#> Iteration: 300 / 500 [ 60%]  (Sampling)
#> Iteration: 350 / 500 [ 70%]  (Sampling)
#> Iteration: 400 / 500 [ 80%]  (Sampling)
#> Iteration: 450 / 500 [ 90%]  (Sampling)
#> Iteration: 500 / 500 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.335691 seconds (Warm-up)
#>                0.15508 seconds (Sampling)
#>                0.490771 seconds (Total)
#> 

example_model

#> stan_glmer
#>  family:   binomial [logit]
#>  formula:  cbind(incidence, size - incidence) ~ size + period + (1 | herd)
#>  num. obs: 56
#> ------
#> 
#> Estimates:
#>             Median MAD_SD
#> (Intercept) -1.5    0.6  
#> size         0.0    0.0  
#> period2     -0.9    0.3  
#> period3     -1.1    0.4  
#> period4     -1.5    0.4  
#> 
#> Error terms:
#>  Groups Name        Std.Dev.
#>  herd   (Intercept) 0.81    
#> Num. levels: herd 15 
#> 
#> Sample avg. posterior predictive 
#> distribution of y (X = xbar):
#>          Median MAD_SD
#> mean_PPD 1.8    0.2   
#> 
#> ------
#> For info on the priors used see help('prior_summary.stanreg').