Pointwise log-likelihood matrix

For models fit using MCMC only, the log_lik method returns the \(S\) by \(N\) pointwise log-likelihood matrix, where \(S\) is the size of the posterior sample and \(N\) is the number of data points.

# S3 method for stanreg
log_lik(object, newdata = NULL, offset = NULL, ...)

Arguments

object
A fitted model object returned by one of the rstanarm modeling functions. See stanreg-objects.

newdata
An optional data frame of new data (e.g. holdout data) to use when evaluating the log-likelihood. See the description of newdata for posterior_predict.

offset
A vector of offsets. Only required if newdata is specified and an offset was specified when fitting the model.

...
Currently ignored.

Value

An \(S\) by \(N\) matrix, where \(S\) is the size of the posterior sample and \(N\) is the number of data points.

Examples

 roaches$roach100 <- roaches$roach1 / 100
 fit <- stan_glm(
    y ~ roach100 + treatment + senior,
    offset = log(exposure2),
    data = roaches,
    family = poisson(link = "log"),
    prior = normal(0, 2.5),
    prior_intercept = normal(0, 10),
    iter = 500 # to speed up example
 )

#> 
#> SAMPLING FOR MODEL 'count' NOW (CHAIN 1).
#> 
#> Chain 1, Iteration:   1 / 500 [  0%]  (Warmup)
#> Chain 1, Iteration:  50 / 500 [ 10%]  (Warmup)
#> Chain 1, Iteration: 100 / 500 [ 20%]  (Warmup)
#> Chain 1, Iteration: 150 / 500 [ 30%]  (Warmup)
#> Chain 1, Iteration: 200 / 500 [ 40%]  (Warmup)
#> Chain 1, Iteration: 250 / 500 [ 50%]  (Warmup)
#> Chain 1, Iteration: 251 / 500 [ 50%]  (Sampling)
#> Chain 1, Iteration: 300 / 500 [ 60%]  (Sampling)
#> Chain 1, Iteration: 350 / 500 [ 70%]  (Sampling)
#> Chain 1, Iteration: 400 / 500 [ 80%]  (Sampling)
#> Chain 1, Iteration: 450 / 500 [ 90%]  (Sampling)
#> Chain 1, Iteration: 500 / 500 [100%]  (Sampling)
#>  Elapsed Time: 0.139602 seconds (Warm-up)
#>                0.116159 seconds (Sampling)
#>                0.255761 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'count' NOW (CHAIN 2).
#> 
#> Chain 2, Iteration:   1 / 500 [  0%]  (Warmup)
#> Chain 2, Iteration:  50 / 500 [ 10%]  (Warmup)
#> Chain 2, Iteration: 100 / 500 [ 20%]  (Warmup)
#> Chain 2, Iteration: 150 / 500 [ 30%]  (Warmup)
#> Chain 2, Iteration: 200 / 500 [ 40%]  (Warmup)
#> Chain 2, Iteration: 250 / 500 [ 50%]  (Warmup)
#> Chain 2, Iteration: 251 / 500 [ 50%]  (Sampling)
#> Chain 2, Iteration: 300 / 500 [ 60%]  (Sampling)
#> Chain 2, Iteration: 350 / 500 [ 70%]  (Sampling)
#> Chain 2, Iteration: 400 / 500 [ 80%]  (Sampling)
#> Chain 2, Iteration: 450 / 500 [ 90%]  (Sampling)
#> Chain 2, Iteration: 500 / 500 [100%]  (Sampling)
#>  Elapsed Time: 0.105851 seconds (Warm-up)
#>                0.129603 seconds (Sampling)
#>                0.235454 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'count' NOW (CHAIN 3).
#> 
#> Chain 3, Iteration:   1 / 500 [  0%]  (Warmup)
#> Chain 3, Iteration:  50 / 500 [ 10%]  (Warmup)
#> Chain 3, Iteration: 100 / 500 [ 20%]  (Warmup)
#> Chain 3, Iteration: 150 / 500 [ 30%]  (Warmup)
#> Chain 3, Iteration: 200 / 500 [ 40%]  (Warmup)
#> Chain 3, Iteration: 250 / 500 [ 50%]  (Warmup)
#> Chain 3, Iteration: 251 / 500 [ 50%]  (Sampling)
#> Chain 3, Iteration: 300 / 500 [ 60%]  (Sampling)
#> Chain 3, Iteration: 350 / 500 [ 70%]  (Sampling)
#> Chain 3, Iteration: 400 / 500 [ 80%]  (Sampling)
#> Chain 3, Iteration: 450 / 500 [ 90%]  (Sampling)
#> Chain 3, Iteration: 500 / 500 [100%]  (Sampling)
#>  Elapsed Time: 0.166166 seconds (Warm-up)
#>                0.128214 seconds (Sampling)
#>                0.29438 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'count' NOW (CHAIN 4).
#> 
#> Chain 4, Iteration:   1 / 500 [  0%]  (Warmup)
#> Chain 4, Iteration:  50 / 500 [ 10%]  (Warmup)
#> Chain 4, Iteration: 100 / 500 [ 20%]  (Warmup)
#> Chain 4, Iteration: 150 / 500 [ 30%]  (Warmup)
#> Chain 4, Iteration: 200 / 500 [ 40%]  (Warmup)
#> Chain 4, Iteration: 250 / 500 [ 50%]  (Warmup)
#> Chain 4, Iteration: 251 / 500 [ 50%]  (Sampling)
#> Chain 4, Iteration: 300 / 500 [ 60%]  (Sampling)
#> Chain 4, Iteration: 350 / 500 [ 70%]  (Sampling)
#> Chain 4, Iteration: 400 / 500 [ 80%]  (Sampling)
#> Chain 4, Iteration: 450 / 500 [ 90%]  (Sampling)
#> Chain 4, Iteration: 500 / 500 [100%]  (Sampling)
#>  Elapsed Time: 0.123372 seconds (Warm-up)
#>                0.154062 seconds (Sampling)
#>                0.277434 seconds (Total)
#> 

 ll <- log_lik(fit)
 dim(ll)

#> [1] 1000  262

 all.equal(ncol(ll), nobs(fit))

#> [1] TRUE

 # using newdata argument
 nd <- roaches[1:2, ]
 nd$treatment[1:2] <- c(0, 1)
 ll2 <- log_lik(fit, newdata = nd, offset = c(0, 0))
 head(ll2)

#>              1         2
#> [1,] -7.030876 -3.468675
#> [2,] -6.693435 -3.398425
#> [3,] -6.743831 -3.479912
#> [4,] -8.131682 -3.609733
#> [5,] -6.158471 -3.342550
#> [6,] -5.852850 -3.342724

 dim(ll2)

#> [1] 1000    2

 all.equal(ncol(ll2), nrow(nd))

#> [1] TRUE