Posterior distribution of the linear predictor

Extract the posterior draws of the linear predictor, possibly transformed by the inverse-link function. This function is occasionally useful, but it should be used sparingly. Inference and model checking should generally be carried out using the posterior predictive distribution (i.e., using posterior_predict).

# S3 method for stanreg
posterior_linpred(object, transform = FALSE,
  newdata = NULL, draws = NULL, re.form = NULL, offset = NULL,
  XZ = FALSE, ...)

Arguments

object
A fitted model object returned by one of the rstanarm modeling functions. See stanreg-objects.

transform
Should the linear predictor be transformed using the inverse-link function? The default is FALSE, in which case the untransformed linear predictor is returned.

newdata, draws, re.form, offset
Same as for posterior_predict.

XZ
If TRUE then instead of computing the linear predictor the design matrix X (or cbind(X,Z) for models with group-specific terms) constructed from newdata is returned. The default is FALSE.

...
Currently ignored.

Value

The default is to return a draws by nrow(newdata) matrix of simulations from the posterior distribution of the (possibly transformed) linear predictor. The exception is if the argument XZ is set to TRUE (see the XZ argument description above).

Note

For models estimated with stan_clogit, the number of successes per stratum is ostensibly fixed by the research design. Thus, when calling posterior_linpred with new data and transform = TRUE, the data.frame passed to the newdata argument must contain an outcome variable and a stratifying factor, both with the same name as in the original data.frame. Then, the probabilities will condition on this outcome in the new data.

See also

posterior_predict to draw from the posterior predictive distribution of the outcome, which is typically preferable.

Examples

if (!exists("example_model")) example(example_model)
print(family(example_model))

#> 
#> Family: binomial 
#> Link function: logit 
#> 

# linear predictor on log-odds scale
linpred <- posterior_linpred(example_model)
colMeans(linpred)

#>          1          2          3          4          5          6          7 
#> -0.8267767 -1.8373747 -1.9697531 -2.4828881 -1.7430203 -2.7714209 -2.8503914 
#>          8          9         10         11         12         13         14 
#> -1.0002683 -2.0464715 -2.1521460 -2.5940705 -1.4468394 -2.4396348 -2.5542106 
#>         15         16         17         18         19         20         21 
#> -3.0584443 -1.6610352 -2.5915213 -2.7060971 -3.3616532 -1.9119590 -2.9047544 
#>         22         23         24         25         26         27         28 
#> -3.0015275 -3.4612547 -0.4976494 -1.5438526 -1.6584284 -2.1715635 -0.7749472 
#>         29         30         31         32         33         34         35 
#> -1.7402832 -2.7597825 -2.8476543 -3.3429868 -2.0441779 -3.0369733 -3.1782530 
#>         36         37         38         39         40         41         42 
#> -3.6201775 -1.5320158 -2.5070085 -2.6571896 -3.1347194 -1.5459130 -2.5565109 
#>         43         44         45         46         47         48         49 
#> -2.6799880 -3.1664191 -2.1997031 -3.1657945 -3.3159755 -3.7579000 -0.3628684 
#>         50         51         52         53         54         55         56 
#> -1.5069860 -1.6037592 -2.0901903 -2.0056602 -3.0340608 -3.1397352 -3.6172650 

# probabilities
probs <- posterior_linpred(example_model, transform = TRUE)
colMeans(probs)

#>          1          2          3          4          5          6          7 
#> 0.31014900 0.14610860 0.13031499 0.08680943 0.15624827 0.06396252 0.05962641 
#>          8          9         10         11         12         13         14 
#> 0.27362385 0.12038861 0.10919710 0.07579599 0.20085737 0.08829532 0.07962574 
#>         15         16         17         18         19         20         21 
#> 0.05151852 0.16679352 0.07476684 0.06754332 0.03986671 0.13706251 0.05683409 
#>         22         23         24         25         26         27         28 
#> 0.05203634 0.03460963 0.38181981 0.18478710 0.16927541 0.11330854 0.32032472 
#>         29         30         31         32         33         34         35 
#> 0.16112481 0.06781147 0.06075172 0.03988276 0.12300191 0.05081201 0.04446909 
#>         36         37         38         39         40         41         42 
#> 0.03038804 0.18294856 0.08061644 0.06958894 0.04614921 0.18735490 0.08151150 
#>         43         44         45         46         47         48         49 
#> 0.07168726 0.04718736 0.10787583 0.04507824 0.03919411 0.02689086 0.41380415 
#>         50         51         52         53         54         55         56 
#> 0.19794756 0.18091288 0.12435975 0.12586694 0.05064797 0.04549723 0.03017447 

# not conditioning on any group-level parameters
probs2 <- posterior_linpred(example_model, transform = TRUE, re.form = NA)
apply(probs2, 2, median)

#>          1          2          3          4          5          6          7 
#> 0.19050884 0.07719124 0.06943886 0.04351901 0.20118629 0.08175936 0.07703424 
#>          8          9         10         11         12         13         14 
#> 0.20118629 0.08149972 0.07392713 0.05032924 0.18582591 0.07569139 0.06943886 
#>         15         16         17         18         19         20         21 
#> 0.04424392 0.19625747 0.08660440 0.07958648 0.04293409 0.19528695 0.08185848 
#>         22         23         24         25         26         27         28 
#> 0.07479393 0.05032924 0.19376054 0.07569139 0.06943886 0.04351901 0.21992122 
#>         29         30         31         32         33         34         35 
#> 0.18429890 0.07363089 0.06865920 0.04424392 0.20118629 0.08573012 0.07479393 
#>         36         37         38         39         40         41         42 
#> 0.05072863 0.20526420 0.08860947 0.07781148 0.05072863 0.18582591 0.07467083 
#>         43         44         45         46         47         48         49 
#> 0.06699119 0.04351901 0.20000985 0.08604035 0.07532156 0.05138197 0.19734840 
#>         50         51         52         53         54         55         56 
#> 0.07062181 0.06577282 0.04147431 0.19734840 0.08029068 0.07327571 0.04757747