Posterior distribution of the linear predictor

Extract the posterior draws of the linear predictor, possibly transformed by the inverse-link function. This function is occasionally useful, but it should be used sparingly. Inference and model checking should generally be carried out using the posterior predictive distribution (i.e., using posterior_predict).

# S3 method for stanreg
posterior_linpred(object, transform = FALSE,
  newdata = NULL, draws = NULL, re.form = NULL, offset = NULL,
  XZ = FALSE, ...)

Arguments

object
A fitted model object returned by one of the rstanarm modeling functions. See stanreg-objects.

transform
Should the linear predictor be transformed using the inverse-link function? The default is FALSE, in which case the untransformed linear predictor is returned.

newdata, draws, re.form, offset
Same as for posterior_predict.

XZ
If TRUE then instead of computing the linear predictor the design matrix X (or cbind(X,Z) for models with group-specific terms) constructed from newdata is returned. The default is FALSE.

...
Currently ignored.

Value

The default is to return a draws by nrow(newdata) matrix of simulations from the posterior distribution of the (possibly transformed) linear predictor. The exception is if the argument XZ is set to TRUE (see the XZ argument description above).

Note

For models estimated with stan_clogit, the number of successes per stratum is ostensibly fixed by the research design. Thus, when calling posterior_linpred with new data and transform = TRUE, the data.frame passed to the newdata argument must contain an outcome variable and a stratifying factor, both with the same name as in the original data.frame. Then, the probabilities will condition on this outcome in the new data.

See also

Examples

if (!exists("example_model")) example(example_model)

#> Warning: no help found for ‘example_model’

print(family(example_model))

#> Error in family(example_model): object 'example_model' not found

# linear predictor on log-odds scale
linpred <- posterior_linpred(example_model)

#> Error in posterior_linpred(example_model): object 'example_model' not found

colMeans(linpred)

#> Error in is.data.frame(x): object 'linpred' not found

# probabilities
probs <- posterior_linpred(example_model, transform = TRUE)

#> Error in posterior_linpred(example_model, transform = TRUE): object 'example_model' not found

colMeans(probs)

#> Error in is.data.frame(x): object 'probs' not found

# not conditioning on any group-level parameters
probs2 <- posterior_linpred(example_model, transform = TRUE, re.form = NA)

#> Error in posterior_linpred(example_model, transform = TRUE, re.form = NA): object 'example_model' not found

apply(probs2, 2, median)

#> Error in apply(probs2, 2, median): object 'probs2' not found