Predictive intervals

For models fit using MCMC (algorithm="sampling") or one of the variational approximations ("meanfield" or "fullrank"), the predictive_interval function computes Bayesian predictive intervals. The method for stanreg objects calls posterior_predict internally, whereas the method for objects of class "ppd" accepts the matrix returned by posterior_predict as input and can be used to avoid multiple calls to posterior_predict.

# S3 method for stanreg
predictive_interval(object, prob = 0.9, newdata = NULL,
  draws = NULL, re.form = NULL, fun = NULL, seed = NULL,
  offset = NULL, ...)

# S3 method for ppd
predictive_interval(object, prob = 0.9, ...)

Arguments

object
Either a fitted model object returned by one of the rstanarm modeling functions (a stanreg object) or, for the "ppd" method, a matrix of draws from the posterior predictive distribution returned by posterior_predict.

prob
A number \(p \in (0,1)\) indicating the desired probability mass to include in the intervals. The default is to report \(90\)% intervals (prob=0.9) rather than the traditionally used \(95\)% (see Details).

newdata, draws, fun, offset, re.form, seed
Passed to posterior_predict.

...
Currently ignored.

Value

A matrix with two columns and as many rows as are in newdata. If newdata is not provided then the matrix will have as many rows as the data used to fit the model. For a given value of prob, \(p\), the columns correspond to the lower and upper \(100p\)% central interval limits and have the names \(100\alpha/2\)% and \(100(1 - \alpha/2)\)%, where \(\alpha = 1-p\). For example, if prob=0.9 is specified (a \(90\)% interval), then the column names will be "5%" and "95%", respectively.

See also

predictive_error, posterior_predict, posterior_interval

Examples

fit <- stan_glm(mpg ~ wt, data = mtcars, iter = 300)

#> trying deprecated constructor; please alert package maintainer

#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 1).
#> 
#> Gradient evaluation took 2.5e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.25 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:   1 / 300 [  0%]  (Warmup)
#> Iteration:  30 / 300 [ 10%]  (Warmup)
#> Iteration:  60 / 300 [ 20%]  (Warmup)
#> Iteration:  90 / 300 [ 30%]  (Warmup)
#> Iteration: 120 / 300 [ 40%]  (Warmup)
#> Iteration: 150 / 300 [ 50%]  (Warmup)
#> Iteration: 151 / 300 [ 50%]  (Sampling)
#> Iteration: 180 / 300 [ 60%]  (Sampling)
#> Iteration: 210 / 300 [ 70%]  (Sampling)
#> Iteration: 240 / 300 [ 80%]  (Sampling)
#> Iteration: 270 / 300 [ 90%]  (Sampling)
#> Iteration: 300 / 300 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.031267 seconds (Warm-up)
#>                0.010641 seconds (Sampling)
#>                0.041908 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 2).
#> 
#> Gradient evaluation took 1.8e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.18 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:   1 / 300 [  0%]  (Warmup)
#> Iteration:  30 / 300 [ 10%]  (Warmup)
#> Iteration:  60 / 300 [ 20%]  (Warmup)
#> Iteration:  90 / 300 [ 30%]  (Warmup)
#> Iteration: 120 / 300 [ 40%]  (Warmup)
#> Iteration: 150 / 300 [ 50%]  (Warmup)
#> Iteration: 151 / 300 [ 50%]  (Sampling)
#> Iteration: 180 / 300 [ 60%]  (Sampling)
#> Iteration: 210 / 300 [ 70%]  (Sampling)
#> Iteration: 240 / 300 [ 80%]  (Sampling)
#> Iteration: 270 / 300 [ 90%]  (Sampling)
#> Iteration: 300 / 300 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.029061 seconds (Warm-up)
#>                0.011599 seconds (Sampling)
#>                0.04066 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 3).
#> 
#> Gradient evaluation took 3.3e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.33 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:   1 / 300 [  0%]  (Warmup)
#> Iteration:  30 / 300 [ 10%]  (Warmup)
#> Iteration:  60 / 300 [ 20%]  (Warmup)
#> Iteration:  90 / 300 [ 30%]  (Warmup)
#> Iteration: 120 / 300 [ 40%]  (Warmup)
#> Iteration: 150 / 300 [ 50%]  (Warmup)
#> Iteration: 151 / 300 [ 50%]  (Sampling)
#> Iteration: 180 / 300 [ 60%]  (Sampling)
#> Iteration: 210 / 300 [ 70%]  (Sampling)
#> Iteration: 240 / 300 [ 80%]  (Sampling)
#> Iteration: 270 / 300 [ 90%]  (Sampling)
#> Iteration: 300 / 300 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.030445 seconds (Warm-up)
#>                0.007554 seconds (Sampling)
#>                0.037999 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 4).
#> 
#> Gradient evaluation took 2.4e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.24 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:   1 / 300 [  0%]  (Warmup)
#> Iteration:  30 / 300 [ 10%]  (Warmup)
#> Iteration:  60 / 300 [ 20%]  (Warmup)
#> Iteration:  90 / 300 [ 30%]  (Warmup)
#> Iteration: 120 / 300 [ 40%]  (Warmup)
#> Iteration: 150 / 300 [ 50%]  (Warmup)
#> Iteration: 151 / 300 [ 50%]  (Sampling)
#> Iteration: 180 / 300 [ 60%]  (Sampling)
#> Iteration: 210 / 300 [ 70%]  (Sampling)
#> Iteration: 240 / 300 [ 80%]  (Sampling)
#> Iteration: 270 / 300 [ 90%]  (Sampling)
#> Iteration: 300 / 300 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.017102 seconds (Warm-up)
#>                0.008071 seconds (Sampling)
#>                0.025173 seconds (Total)
#> 

predictive_interval(fit)

#>                            5%      95%
#> Mazda RX4           18.243485 28.97967
#> Mazda RX4 Wag       16.480815 26.97427
#> Datsun 710          19.909912 30.64316
#> Hornet 4 Drive      14.766080 26.05541
#> Hornet Sportabout   13.983066 24.20360
#> Valiant             13.343447 24.21001
#> Duster 360          13.153373 23.55708
#> Merc 240D           15.168880 25.57952
#> Merc 230            15.048141 26.53197
#> Merc 280            13.772423 24.70903
#> Merc 280C           13.673049 24.47098
#> Merc 450SE          10.505083 20.42228
#> Merc 450SL          12.172953 22.30860
#> Merc 450SLC         11.626296 22.85241
#> Cadillac Fleetwood   3.565999 15.30422
#> Lincoln Continental  2.100995 13.95298
#> Chrysler Imperial    3.305724 14.39781
#> Fiat 128            20.181822 30.59120
#> Honda Civic         23.609939 34.34089
#> Toyota Corolla      22.160843 33.32656
#> Toyota Corona       18.684313 30.00222
#> Dodge Challenger    13.309043 24.09148
#> AMC Javelin         13.939583 24.42285
#> Camaro Z28          11.585181 22.55156
#> Pontiac Firebird    11.611114 22.07110
#> Fiat X1-9           21.668641 32.79780
#> Porsche 914-2       20.180832 31.42086
#> Lotus Europa        23.887803 35.43061
#> Ford Pantera L      14.887926 25.64580
#> Ferrari Dino        17.450685 28.01122
#> Maserati Bora       12.855986 23.47998
#> Volvo 142E          17.034733 27.93474

predictive_interval(fit, newdata = data.frame(wt = range(mtcars$wt)),
                    prob = 0.5)

#>         25%      75%
#> 1 26.781075 31.29569
#> 2  6.225066 10.70728

# stanreg vs ppd methods
preds <- posterior_predict(fit, seed = 123)
all.equal(
  predictive_interval(fit, seed = 123),
  predictive_interval(preds)
)

#> [1] TRUE