Predictive intervals

For models fit using MCMC (algorithm="sampling") or one of the variational approximations ("meanfield" or "fullrank"), the predictive_interval function computes Bayesian predictive intervals. The method for stanreg objects calls posterior_predict internally, whereas the method for objects of class "ppd" accepts the matrix returned by posterior_predict as input and can be used to avoid multiple calls to posterior_predict.

# S3 method for stanreg
predictive_interval(object, prob = 0.9, newdata = NULL,
  draws = NULL, re.form = NULL, fun = NULL, seed = NULL,
  offset = NULL, ...)

# S3 method for ppd
predictive_interval(object, prob = 0.9, ...)

Arguments

object
Either a fitted model object returned by one of the rstanarm modeling functions (a stanreg object) or, for the "ppd" method, a matrix of draws from the posterior predictive distribution returned by posterior_predict.

prob
A number \(p \in (0,1)\) indicating the desired probability mass to include in the intervals. The default is to report \(90\)% intervals (prob=0.9) rather than the traditionally used \(95\)% (see Details).

newdata, draws, fun, offset, re.form, seed
Passed to posterior_predict.

...
Currently ignored.

Value

A matrix with two columns and as many rows as are in newdata. If newdata is not provided then the matrix will have as many rows as the data used to fit the model. For a given value of prob, \(p\), the columns correspond to the lower and upper \(100p\)% central interval limits and have the names \(100\alpha/2\)% and \(100(1 - \alpha/2)\)%, where \(\alpha = 1-p\). For example, if prob=0.9 is specified (a \(90\)% interval), then the column names will be "5%" and "95%", respectively.

See also

predictive_error, posterior_predict, posterior_interval

Examples

fit <- stan_glm(mpg ~ wt, data = mtcars, iter = 300)

#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 1).
#> 
#> Gradient evaluation took 3.5e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.35 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:   1 / 300 [  0%]  (Warmup)
#> Iteration:  30 / 300 [ 10%]  (Warmup)
#> Iteration:  60 / 300 [ 20%]  (Warmup)
#> Iteration:  90 / 300 [ 30%]  (Warmup)
#> Iteration: 120 / 300 [ 40%]  (Warmup)
#> Iteration: 150 / 300 [ 50%]  (Warmup)
#> Iteration: 151 / 300 [ 50%]  (Sampling)
#> Iteration: 180 / 300 [ 60%]  (Sampling)
#> Iteration: 210 / 300 [ 70%]  (Sampling)
#> Iteration: 240 / 300 [ 80%]  (Sampling)
#> Iteration: 270 / 300 [ 90%]  (Sampling)
#> Iteration: 300 / 300 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.018574 seconds (Warm-up)
#>                0.009285 seconds (Sampling)
#>                0.027859 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 2).
#> 
#> Gradient evaluation took 1.4e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.14 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:   1 / 300 [  0%]  (Warmup)
#> Iteration:  30 / 300 [ 10%]  (Warmup)
#> Iteration:  60 / 300 [ 20%]  (Warmup)
#> Iteration:  90 / 300 [ 30%]  (Warmup)
#> Iteration: 120 / 300 [ 40%]  (Warmup)
#> Iteration: 150 / 300 [ 50%]  (Warmup)
#> Iteration: 151 / 300 [ 50%]  (Sampling)
#> Iteration: 180 / 300 [ 60%]  (Sampling)
#> Iteration: 210 / 300 [ 70%]  (Sampling)
#> Iteration: 240 / 300 [ 80%]  (Sampling)
#> Iteration: 270 / 300 [ 90%]  (Sampling)
#> Iteration: 300 / 300 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.032385 seconds (Warm-up)
#>                0.007527 seconds (Sampling)
#>                0.039912 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 3).
#> 
#> Gradient evaluation took 1.2e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.12 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:   1 / 300 [  0%]  (Warmup)
#> Iteration:  30 / 300 [ 10%]  (Warmup)
#> Iteration:  60 / 300 [ 20%]  (Warmup)
#> Iteration:  90 / 300 [ 30%]  (Warmup)
#> Iteration: 120 / 300 [ 40%]  (Warmup)
#> Iteration: 150 / 300 [ 50%]  (Warmup)
#> Iteration: 151 / 300 [ 50%]  (Sampling)
#> Iteration: 180 / 300 [ 60%]  (Sampling)
#> Iteration: 210 / 300 [ 70%]  (Sampling)
#> Iteration: 240 / 300 [ 80%]  (Sampling)
#> Iteration: 270 / 300 [ 90%]  (Sampling)
#> Iteration: 300 / 300 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.014375 seconds (Warm-up)
#>                0.007507 seconds (Sampling)
#>                0.021882 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 4).
#> 
#> Gradient evaluation took 2.1e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.21 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:   1 / 300 [  0%]  (Warmup)
#> Iteration:  30 / 300 [ 10%]  (Warmup)
#> Iteration:  60 / 300 [ 20%]  (Warmup)
#> Iteration:  90 / 300 [ 30%]  (Warmup)
#> Iteration: 120 / 300 [ 40%]  (Warmup)
#> Iteration: 150 / 300 [ 50%]  (Warmup)
#> Iteration: 151 / 300 [ 50%]  (Sampling)
#> Iteration: 180 / 300 [ 60%]  (Sampling)
#> Iteration: 210 / 300 [ 70%]  (Sampling)
#> Iteration: 240 / 300 [ 80%]  (Sampling)
#> Iteration: 270 / 300 [ 90%]  (Sampling)
#> Iteration: 300 / 300 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.021222 seconds (Warm-up)
#>                0.009793 seconds (Sampling)
#>                0.031015 seconds (Total)
#> 

predictive_interval(fit)

#>                            5%      95%
#> Mazda RX4           18.512787 28.96880
#> Mazda RX4 Wag       16.928403 26.93170
#> Datsun 710          19.746582 30.30137
#> Hornet 4 Drive      14.467984 25.98264
#> Hornet Sportabout   13.846382 23.78462
#> Valiant             13.255237 23.82880
#> Duster 360          13.212060 23.07730
#> Merc 240D           15.182423 25.78017
#> Merc 230            15.019958 26.19685
#> Merc 280            13.726131 24.76865
#> Merc 280C           13.402244 24.74714
#> Merc 450SE          10.320648 20.40841
#> Merc 450SL          12.152080 22.56150
#> Merc 450SLC         11.953933 22.54094
#> Cadillac Fleetwood   3.757832 15.14776
#> Lincoln Continental  2.507392 14.13168
#> Chrysler Imperial    3.209279 14.50335
#> Fiat 128            20.339031 30.74044
#> Honda Civic         23.732562 34.57353
#> Toyota Corolla      22.025432 32.82623
#> Toyota Corona       18.738266 29.50700
#> Dodge Challenger    12.967483 23.63882
#> AMC Javelin         13.627750 24.29827
#> Camaro Z28          11.494428 22.23448
#> Pontiac Firebird    11.489765 21.79951
#> Fiat X1-9           21.794996 32.60149
#> Porsche 914-2       20.699210 31.23747
#> Lotus Europa        23.845596 35.42853
#> Ford Pantera L      14.830341 25.44669
#> Ferrari Dino        17.484422 27.94563
#> Maserati Bora       12.450753 23.23743
#> Volvo 142E          17.385015 27.80787

predictive_interval(fit, newdata = data.frame(wt = range(mtcars$wt)),
                    prob = 0.5)

#>         25%      75%
#> 1 26.846494 31.06238
#> 2  6.155986 10.56783

# stanreg vs ppd methods
preds <- posterior_predict(fit, seed = 123)
all.equal(
  predictive_interval(fit, seed = 123),
  predictive_interval(preds)
)

#> [1] TRUE