Predictive intervals

For models fit using MCMC (algorithm="sampling") or one of the variational approximations ("meanfield" or "fullrank"), the predictive_interval function computes Bayesian predictive intervals. The method for stanreg objects calls posterior_predict internally, whereas the method for objects of class "ppd" accepts the matrix returned by posterior_predict as input and can be used to avoid multiple calls to posterior_predict.

# S3 method for stanreg
predictive_interval(object, prob = 0.9, newdata = NULL,
  draws = NULL, re.form = NULL, fun = NULL, seed = NULL,
  offset = NULL, ...)

# S3 method for ppd
predictive_interval(object, prob = 0.9, ...)

Arguments

object
Either a fitted model object returned by one of the rstanarm modeling functions (a stanreg object) or, for the "ppd" method, a matrix of draws from the posterior predictive distribution returned by posterior_predict.

prob
A number \(p \in (0,1)\) indicating the desired probability mass to include in the intervals. The default is to report \(90\)% intervals (prob=0.9) rather than the traditionally used \(95\)% (see Details).

newdata, draws, fun, offset, re.form, seed
Passed to posterior_predict.

...
Currently ignored.

Value

A matrix with two columns and as many rows as are in newdata. If newdata is not provided then the matrix will have as many rows as the data used to fit the model. For a given value of prob, \(p\), the columns correspond to the lower and upper \(100p\)% central interval limits and have the names \(100\alpha/2\)% and \(100(1 - \alpha/2)\)%, where \(\alpha = 1-p\). For example, if prob=0.9 is specified (a \(90\)% interval), then the column names will be "5%" and "95%", respectively.

See also

predictive_error, posterior_predict, posterior_interval

Examples

fit <- stan_glm(mpg ~ wt, data = mtcars, iter = 300)

#> trying deprecated constructor; please alert package maintainer

#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 1).
#> 
#> Gradient evaluation took 2.4e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.24 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:   1 / 300 [  0%]  (Warmup)
#> Iteration:  30 / 300 [ 10%]  (Warmup)
#> Iteration:  60 / 300 [ 20%]  (Warmup)
#> Iteration:  90 / 300 [ 30%]  (Warmup)
#> Iteration: 120 / 300 [ 40%]  (Warmup)
#> Iteration: 150 / 300 [ 50%]  (Warmup)
#> Iteration: 151 / 300 [ 50%]  (Sampling)
#> Iteration: 180 / 300 [ 60%]  (Sampling)
#> Iteration: 210 / 300 [ 70%]  (Sampling)
#> Iteration: 240 / 300 [ 80%]  (Sampling)
#> Iteration: 270 / 300 [ 90%]  (Sampling)
#> Iteration: 300 / 300 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.018347 seconds (Warm-up)
#>                0.009595 seconds (Sampling)
#>                0.027942 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 2).
#> 
#> Gradient evaluation took 1.3e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.13 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:   1 / 300 [  0%]  (Warmup)
#> Iteration:  30 / 300 [ 10%]  (Warmup)
#> Iteration:  60 / 300 [ 20%]  (Warmup)
#> Iteration:  90 / 300 [ 30%]  (Warmup)
#> Iteration: 120 / 300 [ 40%]  (Warmup)
#> Iteration: 150 / 300 [ 50%]  (Warmup)
#> Iteration: 151 / 300 [ 50%]  (Sampling)
#> Iteration: 180 / 300 [ 60%]  (Sampling)
#> Iteration: 210 / 300 [ 70%]  (Sampling)
#> Iteration: 240 / 300 [ 80%]  (Sampling)
#> Iteration: 270 / 300 [ 90%]  (Sampling)
#> Iteration: 300 / 300 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.019889 seconds (Warm-up)
#>                0.00848 seconds (Sampling)
#>                0.028369 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 3).
#> 
#> Gradient evaluation took 1.6e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.16 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:   1 / 300 [  0%]  (Warmup)
#> Iteration:  30 / 300 [ 10%]  (Warmup)
#> Iteration:  60 / 300 [ 20%]  (Warmup)
#> Iteration:  90 / 300 [ 30%]  (Warmup)
#> Iteration: 120 / 300 [ 40%]  (Warmup)
#> Iteration: 150 / 300 [ 50%]  (Warmup)
#> Iteration: 151 / 300 [ 50%]  (Sampling)
#> Iteration: 180 / 300 [ 60%]  (Sampling)
#> Iteration: 210 / 300 [ 70%]  (Sampling)
#> Iteration: 240 / 300 [ 80%]  (Sampling)
#> Iteration: 270 / 300 [ 90%]  (Sampling)
#> Iteration: 300 / 300 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.015005 seconds (Warm-up)
#>                0.00736 seconds (Sampling)
#>                0.022365 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 4).
#> 
#> Gradient evaluation took 1.5e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.15 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:   1 / 300 [  0%]  (Warmup)
#> Iteration:  30 / 300 [ 10%]  (Warmup)
#> Iteration:  60 / 300 [ 20%]  (Warmup)
#> Iteration:  90 / 300 [ 30%]  (Warmup)
#> Iteration: 120 / 300 [ 40%]  (Warmup)
#> Iteration: 150 / 300 [ 50%]  (Warmup)
#> Iteration: 151 / 300 [ 50%]  (Sampling)
#> Iteration: 180 / 300 [ 60%]  (Sampling)
#> Iteration: 210 / 300 [ 70%]  (Sampling)
#> Iteration: 240 / 300 [ 80%]  (Sampling)
#> Iteration: 270 / 300 [ 90%]  (Sampling)
#> Iteration: 300 / 300 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.016849 seconds (Warm-up)
#>                0.007486 seconds (Sampling)
#>                0.024335 seconds (Total)
#> 

predictive_interval(fit)

#>                            5%      95%
#> Mazda RX4           18.484634 29.02546
#> Mazda RX4 Wag       16.687102 27.05837
#> Datsun 710          19.592277 30.37484
#> Hornet 4 Drive      14.328101 25.87628
#> Hornet Sportabout   13.461231 23.96949
#> Valiant             13.182912 23.82432
#> Duster 360          13.076391 23.28057
#> Merc 240D           15.079126 25.61700
#> Merc 230            15.229267 26.24887
#> Merc 280            13.869871 25.11116
#> Merc 280C           13.251694 24.43255
#> Merc 450SE          10.287650 20.47020
#> Merc 450SL          12.184462 22.55922
#> Merc 450SLC         11.595785 22.52058
#> Cadillac Fleetwood   3.882155 15.52670
#> Lincoln Continental  2.423747 14.01999
#> Chrysler Imperial    3.062234 14.45498
#> Fiat 128            19.980282 30.59668
#> Honda Civic         23.471849 34.14239
#> Toyota Corolla      21.967879 33.07861
#> Toyota Corona       18.738518 29.78627
#> Dodge Challenger    13.120801 23.92288
#> AMC Javelin         13.968470 24.57885
#> Camaro Z28          11.080500 22.22849
#> Pontiac Firebird    11.345720 21.98579
#> Fiat X1-9           21.605394 32.47980
#> Porsche 914-2       20.361188 30.95626
#> Lotus Europa        23.791260 35.10855
#> Ford Pantera L      15.113103 25.78248
#> Ferrari Dino        17.415401 27.91878
#> Maserati Bora       12.600273 23.36783
#> Volvo 142E          17.240173 27.85937

predictive_interval(fit, newdata = data.frame(wt = range(mtcars$wt)),
                    prob = 0.5)

#>         25%      75%
#> 1 26.627901 31.24743
#> 2  6.426857 10.67194

# stanreg vs ppd methods
preds <- posterior_predict(fit, seed = 123)
all.equal(
  predictive_interval(fit, seed = 123),
  predictive_interval(preds)
)

#> [1] TRUE