Predictive intervals

For models fit using MCMC (algorithm="sampling") or one of the variational approximations ("meanfield" or "fullrank"), the predictive_interval function computes Bayesian predictive intervals. The method for stanreg objects calls posterior_predict internally, whereas the method for objects of class "ppd" accepts the matrix returned by posterior_predict as input and can be used to avoid multiple calls to posterior_predict.

# S3 method for stanreg
predictive_interval(object, prob = 0.9, newdata = NULL,
  draws = NULL, re.form = NULL, fun = NULL, seed = NULL,
  offset = NULL, ...)

# S3 method for ppd
predictive_interval(object, prob = 0.9, ...)

Arguments

object
Either a fitted model object returned by one of the rstanarm modeling functions (a stanreg object) or, for the "ppd" method, a matrix of draws from the posterior predictive distribution returned by posterior_predict.

prob
A number \(p \in (0,1)\) indicating the desired probability mass to include in the intervals. The default is to report \(90\)% intervals (prob=0.9) rather than the traditionally used \(95\)% (see Details).

newdata, draws, fun, offset, re.form, seed
Passed to posterior_predict.

...
Currently ignored.

Value

A matrix with two columns and as many rows as are in newdata. If newdata is not provided then the matrix will have as many rows as the data used to fit the model. For a given value of prob, \(p\), the columns correspond to the lower and upper \(100p\)% central interval limits and have the names \(100\alpha/2\)% and \(100(1 - \alpha/2)\)%, where \(\alpha = 1-p\). For example, if prob=0.9 is specified (a \(90\)% interval), then the column names will be "5%" and "95%", respectively.

See also

predictive_error, posterior_predict, posterior_interval

Examples

fit <- stan_glm(mpg ~ wt, data = mtcars, iter = 300)

#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 1).
#> 
#> Chain 1, Iteration:   1 / 300 [  0%]  (Warmup)
#> Chain 1, Iteration:  30 / 300 [ 10%]  (Warmup)
#> Chain 1, Iteration:  60 / 300 [ 20%]  (Warmup)
#> Chain 1, Iteration:  90 / 300 [ 30%]  (Warmup)
#> Chain 1, Iteration: 120 / 300 [ 40%]  (Warmup)
#> Chain 1, Iteration: 150 / 300 [ 50%]  (Warmup)
#> Chain 1, Iteration: 151 / 300 [ 50%]  (Sampling)
#> Chain 1, Iteration: 180 / 300 [ 60%]  (Sampling)
#> Chain 1, Iteration: 210 / 300 [ 70%]  (Sampling)
#> Chain 1, Iteration: 240 / 300 [ 80%]  (Sampling)
#> Chain 1, Iteration: 270 / 300 [ 90%]  (Sampling)
#> Chain 1, Iteration: 300 / 300 [100%]  (Sampling)
#>  Elapsed Time: 0.021315 seconds (Warm-up)
#>                0.007462 seconds (Sampling)
#>                0.028777 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 2).
#> 
#> Chain 2, Iteration:   1 / 300 [  0%]  (Warmup)
#> Chain 2, Iteration:  30 / 300 [ 10%]  (Warmup)
#> Chain 2, Iteration:  60 / 300 [ 20%]  (Warmup)
#> Chain 2, Iteration:  90 / 300 [ 30%]  (Warmup)
#> Chain 2, Iteration: 120 / 300 [ 40%]  (Warmup)
#> Chain 2, Iteration: 150 / 300 [ 50%]  (Warmup)
#> Chain 2, Iteration: 151 / 300 [ 50%]  (Sampling)
#> Chain 2, Iteration: 180 / 300 [ 60%]  (Sampling)
#> Chain 2, Iteration: 210 / 300 [ 70%]  (Sampling)
#> Chain 2, Iteration: 240 / 300 [ 80%]  (Sampling)
#> Chain 2, Iteration: 270 / 300 [ 90%]  (Sampling)
#> Chain 2, Iteration: 300 / 300 [100%]  (Sampling)
#>  Elapsed Time: 0.016133 seconds (Warm-up)
#>                0.008911 seconds (Sampling)
#>                0.025044 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 3).
#> 
#> Chain 3, Iteration:   1 / 300 [  0%]  (Warmup)
#> Chain 3, Iteration:  30 / 300 [ 10%]  (Warmup)
#> Chain 3, Iteration:  60 / 300 [ 20%]  (Warmup)
#> Chain 3, Iteration:  90 / 300 [ 30%]  (Warmup)
#> Chain 3, Iteration: 120 / 300 [ 40%]  (Warmup)
#> Chain 3, Iteration: 150 / 300 [ 50%]  (Warmup)
#> Chain 3, Iteration: 151 / 300 [ 50%]  (Sampling)
#> Chain 3, Iteration: 180 / 300 [ 60%]  (Sampling)
#> Chain 3, Iteration: 210 / 300 [ 70%]  (Sampling)
#> Chain 3, Iteration: 240 / 300 [ 80%]  (Sampling)
#> Chain 3, Iteration: 270 / 300 [ 90%]  (Sampling)
#> Chain 3, Iteration: 300 / 300 [100%]  (Sampling)
#>  Elapsed Time: 0.015468 seconds (Warm-up)
#>                0.006832 seconds (Sampling)
#>                0.0223 seconds (Total)
#> 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 4).
#> 
#> Chain 4, Iteration:   1 / 300 [  0%]  (Warmup)
#> Chain 4, Iteration:  30 / 300 [ 10%]  (Warmup)
#> Chain 4, Iteration:  60 / 300 [ 20%]  (Warmup)
#> Chain 4, Iteration:  90 / 300 [ 30%]  (Warmup)
#> Chain 4, Iteration: 120 / 300 [ 40%]  (Warmup)
#> Chain 4, Iteration: 150 / 300 [ 50%]  (Warmup)
#> Chain 4, Iteration: 151 / 300 [ 50%]  (Sampling)
#> Chain 4, Iteration: 180 / 300 [ 60%]  (Sampling)
#> Chain 4, Iteration: 210 / 300 [ 70%]  (Sampling)
#> Chain 4, Iteration: 240 / 300 [ 80%]  (Sampling)
#> Chain 4, Iteration: 270 / 300 [ 90%]  (Sampling)
#> Chain 4, Iteration: 300 / 300 [100%]  (Sampling)
#>  Elapsed Time: 0.017651 seconds (Warm-up)
#>                0.006638 seconds (Sampling)
#>                0.024289 seconds (Total)
#> 

predictive_interval(fit)

#>                            5%      95%
#> Mazda RX4           18.366267 28.78491
#> Mazda RX4 Wag       16.308554 27.03948
#> Datsun 710          19.718759 30.21674
#> Hornet 4 Drive      14.648720 25.86737
#> Hornet Sportabout   13.800566 24.26061
#> Valiant             13.571167 24.21148
#> Duster 360          13.445467 23.23589
#> Merc 240D           15.141341 25.61888
#> Merc 230            15.035253 26.61805
#> Merc 280            13.840674 24.89885
#> Merc 280C           13.362155 24.54728
#> Merc 450SE          10.118990 20.31257
#> Merc 450SL          11.669390 22.89848
#> Merc 450SLC         11.717262 22.41068
#> Cadillac Fleetwood   3.765952 15.14795
#> Lincoln Continental  2.652415 14.21044
#> Chrysler Imperial    2.984204 14.37833
#> Fiat 128            19.865395 30.74190
#> Honda Civic         23.360169 34.63433
#> Toyota Corolla      21.771495 33.08279
#> Toyota Corona       18.594855 29.75912
#> Dodge Challenger    12.979858 23.65048
#> AMC Javelin         13.996484 24.21028
#> Camaro Z28          10.976021 22.14573
#> Pontiac Firebird    11.305823 21.90806
#> Fiat X1-9           21.543549 32.69983
#> Porsche 914-2       20.397199 31.33724
#> Lotus Europa        23.786443 35.26948
#> Ford Pantera L      14.751207 25.61128
#> Ferrari Dino        17.481078 27.99494
#> Maserati Bora       12.767369 23.43407
#> Volvo 142E          17.132215 27.44806

predictive_interval(fit, newdata = data.frame(wt = range(mtcars$wt)),
                    prob = 0.5)

#>        25%      75%
#> 1 26.67755 31.33654
#> 2  6.41331 10.83816

# stanreg vs ppd methods
preds <- posterior_predict(fit, seed = 123)
all.equal(
  predictive_interval(fit, seed = 123),
  predictive_interval(preds)
)

#> [1] TRUE