print.stanreg {rstanarm}	R Documentation

Print method for stanreg objects

Description

The print method for stanreg objects displays a compact summary of the fitted model. See the Details section below for a description of the printed output. For additional summary statistics and diagnostics use the summary method.

Usage

## S3 method for class 'stanreg'
print(x, digits = 1, ...)

## S3 method for class 'stanmvreg'
print(x, digits = 3, ...)

Arguments

x
A fitted model object returned by one of the rstanarm modeling functions. See stanreg-objects.

digits
Number of digits to use for formatting numbers.

...
Ignored.

Details

Point estimates

Regardless of the estimation algorithm, point estimates are medians computed from simulations. For models fit using MCMC ("sampling") the posterior sample is used. For optimization ("optimizing"), the simulations are generated from the asymptotic Gaussian sampling distribution of the parameters. For the "meanfield" and "fullrank" variational approximations, draws from the variational approximation to the posterior are used. In all cases, the point estimates reported are the same as the values returned by coef.

Uncertainty estimates

The standard deviations reported (labeled MAD_SD in the print output) are computed from the same set of draws described above and are proportional to the median absolute deviation (mad) from the median. Compared to the raw posterior standard deviation, the MAD_SD will be more robust for long-tailed distributions. These are the same as the values returned by se.

Additional output

For models fit using MCMC or a variational approximation, the median and MAD_SD are also reported for mean_PPD, the sample average posterior predictive distribution of the outcome.

For GLMs with group-specific terms (see stan_glmer) the printed output also shows point estimates of the standard deviations of the group effects (and correlations if there are both intercept and slopes that vary by group).

For analysis of variance models (see stan_aov) models, an ANOVA-like table is also displayed.

For joint longitudinal and time-to-event (see stan_jm) models the estimates are presented separately for each of the distinct submodels.

Value

Returns x, invisibly.

See Also

summary.stanreg, stanreg-methods

[Package rstanarm version 2.17.3 Index]