Methods for stanreg objects

The methods documented on this page are actually some of the least important methods defined for stanreg objects. The most important methods are documented separately, each with its own page. Links to those pages are provided in the See Also section, below.

# S3 method for stanreg
coef(object, ...)

# S3 method for stanreg
confint(object, parm, level = 0.95, ...)

# S3 method for stanreg
fitted(object, ...)

# S3 method for stanreg
nobs(object, ...)

# S3 method for stanreg
residuals(object, ...)

# S3 method for stanreg
se(object, ...)

# S3 method for stanreg
update(object, formula., ..., evaluate = TRUE)

# S3 method for stanreg
vcov(object, correlation = FALSE, ...)

# S3 method for stanreg
fixef(object, ...)

# S3 method for stanreg
ngrps(object, ...)

# S3 method for stanreg
ranef(object, ...)

# S3 method for stanreg
sigma(object, ...)

# S3 method for stanreg
VarCorr(x, sigma = 1, ...)

Arguments

object, x
A fitted model object returned by one of the rstanarm modeling functions. See stanreg-objects.

...
Ignored, except by the update method. See update.

parm
For confint, an optional character vector of parameter names.

level
For confint, a scalar between \(0\) and \(1\) indicating the confidence level to use.

formula., evaluate
See update.

correlation
For vcov, if FALSE (the default) the covariance matrix is returned. If TRUE, the correlation matrix is returned instead.

sigma
Ignored (included for compatibility with VarCorr).

Details

The methods documented on this page are similar to the methods defined for objects of class 'lm', 'glm', 'glmer', etc. However there are a few key differences:

residuals

Residuals are always of type "response" (not "deviance" residuals or any other type). However, in the case of stan_polr with more than two response categories, the residuals are the difference between the latent utility and its linear predictor.

coef

Medians are used for point estimates. See the Point estimates section in print.stanreg for more details.

se

The se function returns standard errors based on mad. See the Uncertainty estimates section in print.stanreg for more details.

confint

For models fit using optimization, confidence intervals are returned via a call to confint.default. If algorithm is "sampling", "meanfield", or "fullrank", the confint will throw an error because the posterior_interval function should be used to compute Bayesian uncertainty intervals.

See also

The print, summary, and prior_summary methods for stanreg objects for information on the fitted model.

launch_shinystan to use the ShinyStan GUI to explore a fitted rstanarm model.

The plot method to plot estimates and diagnostics.

The pp_check method for graphical posterior predictive checking.

The posterior_predict and predictive_error methods for predictions and predictive errors.

The posterior_interval and predictive_interval methods for uncertainty intervals for model parameters and predictions.

The loo, kfold, and log_lik methods for leave-one-out or K-fold cross-validation, model comparison, and computing the log-likelihood of (possibly new) data.

The as.matrix, as.data.frame, and as.array methods to access posterior draws.