Fitted model objects

The rstanarm model-fitting functions return an object of class 'stanreg', which is a list containing at a minimum the components listed below. Each stanreg object will also have additional classes (e.g. 'aov', 'betareg', 'glm', 'polr', etc.) and several additional components depending on the model and estimation algorithm.

Note

The stan_biglm function is an exception. It returns a stanfit object rather than a stanreg object.

stanreg objects

coefficients

Point estimates, as described in print.stanreg.

ses

Standard errors based on mad, as described in print.stanreg.

residuals

Residuals of type 'response'.

fitted.values

Fitted mean values. For GLMs the linear predictors are transformed by the inverse link function.

linear.predictors

Linear fit on the link scale. For linear models this is the same as fitted.values.

covmat

Variance-covariance matrix for the coefficients based on draws from the posterior distribution, the variational approximation, or the asymptotic sampling distribution, depending on the estimation algorithm.

model,x,y

If requested, the the model frame, model matrix and response variable used, respectively.

family

The family object used.

call

The matched call.

formula

The model formula.

data,offset,weights

The data, offset, and weights arguments.

algorithm

The estimation method used.

prior.info

A list with information about the prior distributions used.

stanfit,stan_summary

The object of stanfit-class returned by RStan and a matrix of various summary statistics from the stanfit object.

rstan_version

The version of the rstan package that was used to fit the model.

See also

stanreg-methods