Summary method for stanreg objects

Summaries of parameter estimates and MCMC convergence diagnostics (Monte Carlo error, effective sample size, Rhat).

# S3 method for stanreg
summary(object, pars = NULL, regex_pars = NULL,
  probs = NULL, ..., digits = 1)

# S3 method for summary.stanreg
print(x, digits = max(1, attr(x, "print.digits")),
  ...)

# S3 method for summary.stanreg
as.data.frame(x, ...)

# S3 method for stanmvreg
summary(object, pars = NULL, regex_pars = NULL,
  probs = NULL, ..., digits = 3)

# S3 method for summary.stanmvreg
print(x, digits = max(1, attr(x, "print.digits")),
  ...)

Arguments

object
A fitted model object returned by one of the rstanarm modeling functions. See stanreg-objects.

pars
An optional character vector specifying a subset of parameters to display. Parameters can be specified by name or several shortcuts can be used. Using pars="beta" will restrict the displayed parameters to only the regression coefficients (without the intercept). "alpha" can also be used as a shortcut for "(Intercept)". If the model has varying intercepts and/or slopes they can be selected using pars = "varying".

In addition, for stanmvreg objects there are some additional shortcuts available. Using pars = "long" will display the parameter estimates for the longitudinal submodels only (excluding group-specific pparameters, but including auxiliary parameters). Using pars = "event" will display the parameter estimates for the event submodel only, including any association parameters. Using pars = "assoc" will display only the association parameters. Using pars = "fixef" will display all fixed effects, but not the random effects or the auxiliary parameters. pars and regex_pars are set to NULL then all fixed effect regression coefficients are selected, as well as any auxiliary parameters and the log posterior.

If pars is NULL all parameters are selected for a stanreg object, while for a stanmvreg object all fixed effect regression coefficients are selected as well as any auxiliary parameters and the log posterior. See Examples.

regex_pars
An optional character vector of regular expressions to use for parameter selection. regex_pars can be used in place of pars or in addition to pars. Currently, all functions that accept a regex_pars argument ignore it for models fit using optimization.

probs
For models fit using MCMC or one of the variational algorithms, an optional numeric vector of probabilities passed to quantile.

...
Currently ignored.

digits
Number of digits to use for formatting numbers when printing. When calling summary, the value of digits is stored as the "print.digits" attribute of the returned object.

x
An object of class "summary.stanreg".

Value

The summary method returns an object of class "summary.stanreg" (or "summary.stanmvreg", inheriting "summary.stanreg"), which is a matrix of summary statistics and diagnostics, with attributes storing information for use by the print method. The print method for summary.stanreg or summary.stanmvreg objects is called for its side effect and just returns its input. The as.data.frame method for summary.stanreg objects converts the matrix to a data.frame, preserving row and column names but dropping the print-related attributes.

See also

prior_summary to extract or print a summary of the priors used for a particular model.

Examples

if (!exists("example_model")) example(example_model)
summary(example_model, probs = c(0.1, 0.9))

#> 
#> Model Info:
#> 
#>  function:     stan_glmer
#>  family:       binomial [logit]
#>  formula:      cbind(incidence, size - incidence) ~ size + period + (1 | herd)
#>  algorithm:    sampling
#>  priors:       see help('prior_summary')
#>  sample:       500 (posterior sample size)
#>  observations: 56
#>  groups:       herd (15)
#> 
#> Estimates:
#>                                       mean   sd     10%    90% 
#> (Intercept)                           -1.6    0.6   -2.3   -0.9
#> size                                   0.0    0.0    0.0    0.0
#> period2                               -1.0    0.3   -1.4   -0.6
#> period3                               -1.1    0.3   -1.5   -0.7
#> period4                               -1.6    0.4   -2.1   -1.0
#> b[(Intercept) herd:1]                  0.6    0.4    0.1    1.2
#> b[(Intercept) herd:2]                 -0.4    0.4   -0.9    0.2
#> b[(Intercept) herd:3]                  0.4    0.4   -0.1    0.9
#> b[(Intercept) herd:4]                  0.0    0.5   -0.6    0.6
#> b[(Intercept) herd:5]                 -0.2    0.4   -0.7    0.2
#> b[(Intercept) herd:6]                 -0.5    0.5   -1.1    0.1
#> b[(Intercept) herd:7]                  0.9    0.5    0.4    1.5
#> b[(Intercept) herd:8]                  0.5    0.5   -0.1    1.1
#> b[(Intercept) herd:9]                 -0.2    0.5   -0.9    0.4
#> b[(Intercept) herd:10]                -0.7    0.5   -1.3   -0.1
#> b[(Intercept) herd:11]                -0.2    0.4   -0.7    0.3
#> b[(Intercept) herd:12]                -0.1    0.6   -0.8    0.6
#> b[(Intercept) herd:13]                -0.8    0.5   -1.5   -0.2
#> b[(Intercept) herd:14]                 1.0    0.5    0.4    1.6
#> b[(Intercept) herd:15]                -0.6    0.4   -1.2   -0.1
#> Sigma[herd:(Intercept),(Intercept)]    0.6    0.4    0.3    1.2
#> mean_PPD                               1.8    0.2    1.5    2.1
#> log-posterior                       -116.0    4.3 -121.8 -110.8
#> 
#> Diagnostics:
#>                                     mcse Rhat n_eff
#> (Intercept)                         0.0  1.0  161  
#> size                                0.0  1.0  183  
#> period2                             0.0  1.0  500  
#> period3                             0.0  1.0  500  
#> period4                             0.0  1.0  500  
#> b[(Intercept) herd:1]               0.0  1.0  279  
#> b[(Intercept) herd:2]               0.0  1.0  500  
#> b[(Intercept) herd:3]               0.0  1.0  500  
#> b[(Intercept) herd:4]               0.0  1.0  282  
#> b[(Intercept) herd:5]               0.0  1.0  500  
#> b[(Intercept) herd:6]               0.0  1.0  287  
#> b[(Intercept) herd:7]               0.0  1.0  259  
#> b[(Intercept) herd:8]               0.0  1.0  280  
#> b[(Intercept) herd:9]               0.0  1.0  358  
#> b[(Intercept) herd:10]              0.0  1.0  500  
#> b[(Intercept) herd:11]              0.0  1.0  192  
#> b[(Intercept) herd:12]              0.0  1.0  500  
#> b[(Intercept) herd:13]              0.0  1.0  286  
#> b[(Intercept) herd:14]              0.0  1.0  260  
#> b[(Intercept) herd:15]              0.0  1.0  500  
#> Sigma[herd:(Intercept),(Intercept)] 0.0  1.0  136  
#> mean_PPD                            0.0  1.0  500  
#> log-posterior                       0.3  1.0  172  
#> 
#> For each parameter, mcse is Monte Carlo standard error, n_eff is a crude measure of effective sample size, and Rhat is the potential scale reduction factor on split chains (at convergence Rhat=1).

# These produce the same output for this example, 
# but the second method can be used for any model
summary(example_model, pars = c("(Intercept)", "size",
                                paste0("period", 2:4)))

#> 
#> Model Info:
#> 
#>  function:     stan_glmer
#>  family:       binomial [logit]
#>  formula:      cbind(incidence, size - incidence) ~ size + period + (1 | herd)
#>  algorithm:    sampling
#>  priors:       see help('prior_summary')
#>  sample:       500 (posterior sample size)
#>  observations: 56
#>  groups:       herd (15)
#> 
#> Estimates:
#>               mean   sd   2.5%   25%   50%   75%   97.5%
#> (Intercept) -1.6    0.6 -2.7   -2.0  -1.6  -1.2  -0.4   
#> size         0.0    0.0  0.0    0.0   0.0   0.0   0.1   
#> period2     -1.0    0.3 -1.7   -1.2  -1.0  -0.8  -0.4   
#> period3     -1.1    0.3 -1.7   -1.3  -1.1  -0.9  -0.5   
#> period4     -1.6    0.4 -2.5   -1.8  -1.5  -1.3  -0.8   
#> 
#> Diagnostics:
#>             mcse Rhat n_eff
#> (Intercept) 0.0  1.0  161  
#> size        0.0  1.0  183  
#> period2     0.0  1.0  500  
#> period3     0.0  1.0  500  
#> period4     0.0  1.0  500  
#> 
#> For each parameter, mcse is Monte Carlo standard error, n_eff is a crude measure of effective sample size, and Rhat is the potential scale reduction factor on split chains (at convergence Rhat=1).

summary(example_model, pars = c("alpha", "beta"))

#> 
#> Model Info:
#> 
#>  function:     stan_glmer
#>  family:       binomial [logit]
#>  formula:      cbind(incidence, size - incidence) ~ size + period + (1 | herd)
#>  algorithm:    sampling
#>  priors:       see help('prior_summary')
#>  sample:       500 (posterior sample size)
#>  observations: 56
#>  groups:       herd (15)
#> 
#> Estimates:
#>               mean   sd   2.5%   25%   50%   75%   97.5%
#> (Intercept) -1.6    0.6 -2.7   -2.0  -1.6  -1.2  -0.4   
#> size         0.0    0.0  0.0    0.0   0.0   0.0   0.1   
#> period2     -1.0    0.3 -1.7   -1.2  -1.0  -0.8  -0.4   
#> period3     -1.1    0.3 -1.7   -1.3  -1.1  -0.9  -0.5   
#> period4     -1.6    0.4 -2.5   -1.8  -1.5  -1.3  -0.8   
#> 
#> Diagnostics:
#>             mcse Rhat n_eff
#> (Intercept) 0.0  1.0  161  
#> size        0.0  1.0  183  
#> period2     0.0  1.0  500  
#> period3     0.0  1.0  500  
#> period4     0.0  1.0  500  
#> 
#> For each parameter, mcse is Monte Carlo standard error, n_eff is a crude measure of effective sample size, and Rhat is the potential scale reduction factor on split chains (at convergence Rhat=1).

# Only show parameters varying by group
summary(example_model, pars = "varying")

#> 
#> Model Info:
#> 
#>  function:     stan_glmer
#>  family:       binomial [logit]
#>  formula:      cbind(incidence, size - incidence) ~ size + period + (1 | herd)
#>  algorithm:    sampling
#>  priors:       see help('prior_summary')
#>  sample:       500 (posterior sample size)
#>  observations: 56
#>  groups:       herd (15)
#> 
#> Estimates:
#>                          mean   sd   2.5%   25%   50%   75%   97.5%
#> b[(Intercept) herd:1]   0.6    0.4 -0.3    0.3   0.6   0.9   1.4   
#> b[(Intercept) herd:2]  -0.4    0.4 -1.3   -0.6  -0.3   0.0   0.4   
#> b[(Intercept) herd:3]   0.4    0.4 -0.3    0.2   0.4   0.6   1.1   
#> b[(Intercept) herd:4]   0.0    0.5 -1.0   -0.3   0.1   0.3   0.9   
#> b[(Intercept) herd:5]  -0.2    0.4 -1.0   -0.5  -0.2   0.0   0.5   
#> b[(Intercept) herd:6]  -0.5    0.5 -1.4   -0.8  -0.5  -0.2   0.3   
#> b[(Intercept) herd:7]   0.9    0.5  0.0    0.6   1.0   1.2   1.8   
#> b[(Intercept) herd:8]   0.5    0.5 -0.5    0.2   0.5   0.8   1.5   
#> b[(Intercept) herd:9]  -0.2    0.5 -1.3   -0.6  -0.2   0.1   0.7   
#> b[(Intercept) herd:10] -0.7    0.5 -1.7   -1.0  -0.7  -0.3   0.2   
#> b[(Intercept) herd:11] -0.2    0.4 -1.0   -0.5  -0.2   0.1   0.5   
#> b[(Intercept) herd:12] -0.1    0.6 -1.2   -0.4  -0.1   0.3   1.0   
#> b[(Intercept) herd:13] -0.8    0.5 -1.9   -1.1  -0.8  -0.5   0.1   
#> b[(Intercept) herd:14]  1.0    0.5  0.2    0.7   1.0   1.3   2.0   
#> b[(Intercept) herd:15] -0.6    0.4 -1.5   -0.9  -0.6  -0.3   0.2   
#> 
#> Diagnostics:
#>                        mcse Rhat n_eff
#> b[(Intercept) herd:1]  0.0  1.0  279  
#> b[(Intercept) herd:2]  0.0  1.0  500  
#> b[(Intercept) herd:3]  0.0  1.0  500  
#> b[(Intercept) herd:4]  0.0  1.0  282  
#> b[(Intercept) herd:5]  0.0  1.0  500  
#> b[(Intercept) herd:6]  0.0  1.0  287  
#> b[(Intercept) herd:7]  0.0  1.0  259  
#> b[(Intercept) herd:8]  0.0  1.0  280  
#> b[(Intercept) herd:9]  0.0  1.0  358  
#> b[(Intercept) herd:10] 0.0  1.0  500  
#> b[(Intercept) herd:11] 0.0  1.0  192  
#> b[(Intercept) herd:12] 0.0  1.0  500  
#> b[(Intercept) herd:13] 0.0  1.0  286  
#> b[(Intercept) herd:14] 0.0  1.0  260  
#> b[(Intercept) herd:15] 0.0  1.0  500  
#> 
#> For each parameter, mcse is Monte Carlo standard error, n_eff is a crude measure of effective sample size, and Rhat is the potential scale reduction factor on split chains (at convergence Rhat=1).

as.data.frame(summary(example_model, pars = "varying"))

#>                               mean       mcse        sd       2.5%        25%
#> b[(Intercept) herd:1]   0.61988940 0.02607362 0.4356440 -0.2939807  0.3301198
#> b[(Intercept) herd:2]  -0.36756467 0.01991316 0.4452717 -1.3324743 -0.6363567
#> b[(Intercept) herd:3]   0.37518732 0.01635374 0.3656808 -0.3354408  0.1515352
#> b[(Intercept) herd:4]   0.03543192 0.02878589 0.4835499 -1.0013518 -0.2525088
#> b[(Intercept) herd:5]  -0.24997431 0.01767584 0.3952438 -1.0333337 -0.5184788
#> b[(Intercept) herd:6]  -0.49199688 0.02738699 0.4643358 -1.4397634 -0.7768218
#> b[(Intercept) herd:7]   0.93121409 0.02855496 0.4593327  0.0325538  0.6416929
#> b[(Intercept) herd:8]   0.49369268 0.03111687 0.5205895 -0.4752695  0.1519361
#> b[(Intercept) herd:9]  -0.24911053 0.02704733 0.5121083 -1.3053476 -0.5902022
#> b[(Intercept) herd:10] -0.66872231 0.02206535 0.4933963 -1.7131355 -0.9850678
#> b[(Intercept) herd:11] -0.18326417 0.02982111 0.4133192 -1.0328134 -0.4697464
#> b[(Intercept) herd:12] -0.06364160 0.02536641 0.5672102 -1.1850429 -0.3725619
#> b[(Intercept) herd:13] -0.81534615 0.02961565 0.5008674 -1.9431440 -1.0812578
#> b[(Intercept) herd:14]  1.03929117 0.02883678 0.4653802  0.1657049  0.7137461
#> b[(Intercept) herd:15] -0.60350064 0.01988222 0.4445799 -1.5096979 -0.8975426
#>                                50%         75%      97.5% n_eff     Rhat
#> b[(Intercept) herd:1]   0.63289111  0.90448412 1.42638663   279 1.001646
#> b[(Intercept) herd:2]  -0.34302254 -0.04480936 0.42728462   500 1.007917
#> b[(Intercept) herd:3]   0.36912591  0.59922051 1.09668159   500 1.003255
#> b[(Intercept) herd:4]   0.05161144  0.34236933 0.88754860   282 1.001602
#> b[(Intercept) herd:5]  -0.22992328  0.01743121 0.48344710   500 1.012396
#> b[(Intercept) herd:6]  -0.47969855 -0.19283280 0.33710511   287 1.018340
#> b[(Intercept) herd:7]   0.96006368  1.23798324 1.84536011   259 1.009413
#> b[(Intercept) herd:8]   0.48965828  0.82775916 1.52498078   280 1.007357
#> b[(Intercept) herd:9]  -0.21765771  0.07750291 0.72809432   358 1.001657
#> b[(Intercept) herd:10] -0.65477939 -0.31066502 0.21655952   500 1.001620
#> b[(Intercept) herd:11] -0.15783749  0.09534273 0.54101551   192 1.023096
#> b[(Intercept) herd:12] -0.07006598  0.28700916 1.03536587   500 1.008701
#> b[(Intercept) herd:13] -0.78210322 -0.45591670 0.05404102   286 1.010897
#> b[(Intercept) herd:14]  1.03433233  1.34098590 1.95066228   260 1.013840
#> b[(Intercept) herd:15] -0.56355838 -0.29960992 0.23571934   500 1.002984