Summary method for stanreg objects

Summaries of parameter estimates and MCMC convergence diagnostics (Monte Carlo error, effective sample size, Rhat).

# S3 method for stanreg
summary(object, pars = NULL, regex_pars = NULL,
  probs = c(0.1, 0.5, 0.9), ..., digits = 1)

# S3 method for summary.stanreg
print(x, digits = max(1, attr(x,
  "print.digits")), ...)

# S3 method for summary.stanreg
as.data.frame(x, ...)

# S3 method for stanmvreg
summary(object, pars = NULL, regex_pars = NULL,
  probs = NULL, ..., digits = 3)

# S3 method for summary.stanmvreg
print(x, digits = max(1, attr(x,
  "print.digits")), ...)

Arguments

object
A fitted model object returned by one of the rstanarm modeling functions. See stanreg-objects.

pars
An optional character vector specifying a subset of parameters to display. Parameters can be specified by name or several shortcuts can be used. Using pars="beta" will restrict the displayed parameters to only the regression coefficients (without the intercept). "alpha" can also be used as a shortcut for "(Intercept)". If the model has varying intercepts and/or slopes they can be selected using pars = "varying".

In addition, for stanmvreg objects there are some additional shortcuts available. Using pars = "long" will display the parameter estimates for the longitudinal submodels only (excluding group-specific pparameters, but including auxiliary parameters). Using pars = "event" will display the parameter estimates for the event submodel only, including any association parameters. Using pars = "assoc" will display only the association parameters. Using pars = "fixef" will display all fixed effects, but not the random effects or the auxiliary parameters. pars and regex_pars are set to NULL then all fixed effect regression coefficients are selected, as well as any auxiliary parameters and the log posterior.

If pars is NULL all parameters are selected for a stanreg object, while for a stanmvreg object all fixed effect regression coefficients are selected as well as any auxiliary parameters and the log posterior. See Examples.

regex_pars
An optional character vector of regular expressions to use for parameter selection. regex_pars can be used in place of pars or in addition to pars. Currently, all functions that accept a regex_pars argument ignore it for models fit using optimization.

probs
For models fit using MCMC or one of the variational algorithms, an optional numeric vector of probabilities passed to quantile.

...
Currently ignored.

digits
Number of digits to use for formatting numbers when printing. When calling summary, the value of digits is stored as the "print.digits" attribute of the returned object.

x
An object of class "summary.stanreg".

Value

The summary method returns an object of class "summary.stanreg" (or "summary.stanmvreg", inheriting "summary.stanreg"), which is a matrix of summary statistics and diagnostics, with attributes storing information for use by the print method. The print method for summary.stanreg or summary.stanmvreg objects is called for its side effect and just returns its input. The as.data.frame method for summary.stanreg objects converts the matrix to a data.frame, preserving row and column names but dropping the print-related attributes.

Details

mean_PPD diagnostic

Summary statistics are also reported for mean_PPD, the sample average posterior predictive distribution of the outcome. This is useful as a quick diagnostic. A useful heuristic is to check if mean_PPD is plausible when compared to mean(y). If it is plausible then this does not mean that the model is good in general (only that it can reproduce the sample mean), however if mean_PPD is implausible then it is a sign that something is wrong (severe model misspecification, problems with the data, computational issues, etc.).

See also

Examples

if (!exists("example_model")) example(example_model)
summary(example_model, probs = c(0.1, 0.9))

#> 
#> Model Info:
#>  function:     stan_glmer
#>  family:       binomial [logit]
#>  formula:      cbind(incidence, size - incidence) ~ size + period + (1 | herd)
#>  algorithm:    sampling
#>  sample:       500 (posterior sample size)
#>  priors:       see help('prior_summary')
#>  observations: 56
#>  groups:       herd (15)
#> 
#> Estimates:
#>                                       mean   sd   10%   90%
#> (Intercept)                         -1.6    0.6 -2.3  -0.8 
#> size                                 0.0    0.0  0.0   0.0 
#> period2                             -1.0    0.3 -1.3  -0.6 
#> period3                             -1.1    0.3 -1.5  -0.7 
#> period4                             -1.6    0.5 -2.2  -1.0 
#> b[(Intercept) herd:1]                0.6    0.4  0.1   1.2 
#> b[(Intercept) herd:2]               -0.4    0.4 -0.9   0.2 
#> b[(Intercept) herd:3]                0.4    0.4 -0.1   0.8 
#> b[(Intercept) herd:4]                0.0    0.5 -0.5   0.6 
#> b[(Intercept) herd:5]               -0.3    0.4 -0.8   0.3 
#> b[(Intercept) herd:6]               -0.5    0.4 -1.1   0.0 
#> b[(Intercept) herd:7]                0.9    0.4  0.4   1.5 
#> b[(Intercept) herd:8]                0.5    0.5  0.0   1.1 
#> b[(Intercept) herd:9]               -0.3    0.6 -1.1   0.5 
#> b[(Intercept) herd:10]              -0.6    0.5 -1.2  -0.1 
#> b[(Intercept) herd:11]              -0.1    0.4 -0.6   0.3 
#> b[(Intercept) herd:12]              -0.1    0.5 -0.8   0.6 
#> b[(Intercept) herd:13]              -0.8    0.5 -1.5  -0.3 
#> b[(Intercept) herd:14]               1.0    0.5  0.5   1.6 
#> b[(Intercept) herd:15]              -0.6    0.5 -1.2  -0.1 
#> Sigma[herd:(Intercept),(Intercept)]  0.6    0.3  0.2   1.0 
#> 
#> Fit Diagnostics:
#>            mean   sd   10%   90%
#> mean_PPD 1.8    0.2  1.5   2.1  
#> 
#> The mean_ppd is the sample average posterior predictive distribution of the outcome variable (for details see help('summary.stanreg')).
#> 
#> MCMC diagnostics
#>                                     mcse Rhat n_eff
#> (Intercept)                         0.0  1.0  203  
#> size                                0.0  1.0  250  
#> period2                             0.0  1.0  881  
#> period3                             0.0  1.0  569  
#> period4                             0.0  1.0  430  
#> b[(Intercept) herd:1]               0.0  1.0  383  
#> b[(Intercept) herd:2]               0.0  1.0  551  
#> b[(Intercept) herd:3]               0.0  1.0  410  
#> b[(Intercept) herd:4]               0.0  1.0  394  
#> b[(Intercept) herd:5]               0.0  1.0  539  
#> b[(Intercept) herd:6]               0.0  1.0  544  
#> b[(Intercept) herd:7]               0.0  1.0  322  
#> b[(Intercept) herd:8]               0.0  1.0  354  
#> b[(Intercept) herd:9]               0.0  1.0  474  
#> b[(Intercept) herd:10]              0.0  1.0  400  
#> b[(Intercept) herd:11]              0.0  1.0  409  
#> b[(Intercept) herd:12]              0.0  1.0  570  
#> b[(Intercept) herd:13]              0.0  1.0  441  
#> b[(Intercept) herd:14]              0.0  1.0  458  
#> b[(Intercept) herd:15]              0.0  1.0  609  
#> Sigma[herd:(Intercept),(Intercept)] 0.0  1.0  227  
#> mean_PPD                            0.0  1.0  455  
#> log-posterior                       0.3  1.0  150  
#> 
#> For each parameter, mcse is Monte Carlo standard error, n_eff is a crude measure of effective sample size, and Rhat is the potential scale reduction factor on split chains (at convergence Rhat=1).

# These produce the same output for this example, 
# but the second method can be used for any model
summary(example_model, pars = c("(Intercept)", "size",
                                paste0("period", 2:4)))

#> 
#> Model Info:
#>  function:     stan_glmer
#>  family:       binomial [logit]
#>  formula:      cbind(incidence, size - incidence) ~ size + period + (1 | herd)
#>  algorithm:    sampling
#>  sample:       500 (posterior sample size)
#>  priors:       see help('prior_summary')
#>  observations: 56
#>  groups:       herd (15)
#> 
#> Estimates:
#>               mean   sd   10%   50%   90%
#> (Intercept) -1.6    0.6 -2.3  -1.5  -0.8 
#> size         0.0    0.0  0.0   0.0   0.0 
#> period2     -1.0    0.3 -1.3  -1.0  -0.6 
#> period3     -1.1    0.3 -1.5  -1.1  -0.7 
#> period4     -1.6    0.5 -2.2  -1.6  -1.0 
#> 
#> MCMC diagnostics
#>             mcse Rhat n_eff
#> (Intercept) 0.0  1.0  203  
#> size        0.0  1.0  250  
#> period2     0.0  1.0  881  
#> period3     0.0  1.0  569  
#> period4     0.0  1.0  430  
#> 
#> For each parameter, mcse is Monte Carlo standard error, n_eff is a crude measure of effective sample size, and Rhat is the potential scale reduction factor on split chains (at convergence Rhat=1).

summary(example_model, pars = c("alpha", "beta"))

#> 
#> Model Info:
#>  function:     stan_glmer
#>  family:       binomial [logit]
#>  formula:      cbind(incidence, size - incidence) ~ size + period + (1 | herd)
#>  algorithm:    sampling
#>  sample:       500 (posterior sample size)
#>  priors:       see help('prior_summary')
#>  observations: 56
#>  groups:       herd (15)
#> 
#> Estimates:
#>               mean   sd   10%   50%   90%
#> (Intercept) -1.6    0.6 -2.3  -1.5  -0.8 
#> size         0.0    0.0  0.0   0.0   0.0 
#> period2     -1.0    0.3 -1.3  -1.0  -0.6 
#> period3     -1.1    0.3 -1.5  -1.1  -0.7 
#> period4     -1.6    0.5 -2.2  -1.6  -1.0 
#> 
#> MCMC diagnostics
#>             mcse Rhat n_eff
#> (Intercept) 0.0  1.0  203  
#> size        0.0  1.0  250  
#> period2     0.0  1.0  881  
#> period3     0.0  1.0  569  
#> period4     0.0  1.0  430  
#> 
#> For each parameter, mcse is Monte Carlo standard error, n_eff is a crude measure of effective sample size, and Rhat is the potential scale reduction factor on split chains (at convergence Rhat=1).

# Only show parameters varying by group
summary(example_model, pars = "varying")

#> 
#> Model Info:
#>  function:     stan_glmer
#>  family:       binomial [logit]
#>  formula:      cbind(incidence, size - incidence) ~ size + period + (1 | herd)
#>  algorithm:    sampling
#>  sample:       500 (posterior sample size)
#>  priors:       see help('prior_summary')
#>  observations: 56
#>  groups:       herd (15)
#> 
#> Estimates:
#>                          mean   sd   10%   50%   90%
#> b[(Intercept) herd:1]   0.6    0.4  0.1   0.6   1.2 
#> b[(Intercept) herd:2]  -0.4    0.4 -0.9  -0.3   0.2 
#> b[(Intercept) herd:3]   0.4    0.4 -0.1   0.4   0.8 
#> b[(Intercept) herd:4]   0.0    0.5 -0.5   0.0   0.6 
#> b[(Intercept) herd:5]  -0.3    0.4 -0.8  -0.2   0.3 
#> b[(Intercept) herd:6]  -0.5    0.4 -1.1  -0.4   0.0 
#> b[(Intercept) herd:7]   0.9    0.4  0.4   0.9   1.5 
#> b[(Intercept) herd:8]   0.5    0.5  0.0   0.5   1.1 
#> b[(Intercept) herd:9]  -0.3    0.6 -1.1  -0.3   0.5 
#> b[(Intercept) herd:10] -0.6    0.5 -1.2  -0.6  -0.1 
#> b[(Intercept) herd:11] -0.1    0.4 -0.6  -0.1   0.3 
#> b[(Intercept) herd:12] -0.1    0.5 -0.8  -0.1   0.6 
#> b[(Intercept) herd:13] -0.8    0.5 -1.5  -0.8  -0.3 
#> b[(Intercept) herd:14]  1.0    0.5  0.5   1.0   1.6 
#> b[(Intercept) herd:15] -0.6    0.5 -1.2  -0.6  -0.1 
#> 
#> MCMC diagnostics
#>                        mcse Rhat n_eff
#> b[(Intercept) herd:1]  0.0  1.0  383  
#> b[(Intercept) herd:2]  0.0  1.0  551  
#> b[(Intercept) herd:3]  0.0  1.0  410  
#> b[(Intercept) herd:4]  0.0  1.0  394  
#> b[(Intercept) herd:5]  0.0  1.0  539  
#> b[(Intercept) herd:6]  0.0  1.0  544  
#> b[(Intercept) herd:7]  0.0  1.0  322  
#> b[(Intercept) herd:8]  0.0  1.0  354  
#> b[(Intercept) herd:9]  0.0  1.0  474  
#> b[(Intercept) herd:10] 0.0  1.0  400  
#> b[(Intercept) herd:11] 0.0  1.0  409  
#> b[(Intercept) herd:12] 0.0  1.0  570  
#> b[(Intercept) herd:13] 0.0  1.0  441  
#> b[(Intercept) herd:14] 0.0  1.0  458  
#> b[(Intercept) herd:15] 0.0  1.0  609  
#> 
#> For each parameter, mcse is Monte Carlo standard error, n_eff is a crude measure of effective sample size, and Rhat is the potential scale reduction factor on split chains (at convergence Rhat=1).

as.data.frame(summary(example_model, pars = "varying"))

#>                               mean       mcse        sd         10%         50%
#> b[(Intercept) herd:1]   0.62179124 0.02232820 0.4372469  0.05695058  0.59668830
#> b[(Intercept) herd:2]  -0.35964139 0.01792239 0.4205940 -0.90738108 -0.33500973
#> b[(Intercept) herd:3]   0.38889627 0.01759661 0.3563992 -0.08772663  0.38967437
#> b[(Intercept) herd:4]   0.03759993 0.02413744 0.4793584 -0.53741816  0.02781152
#> b[(Intercept) herd:5]  -0.25891671 0.01838653 0.4269887 -0.79368970 -0.24284159
#> b[(Intercept) herd:6]  -0.46760700 0.01785554 0.4163340 -1.05385969 -0.44309550
#> b[(Intercept) herd:7]   0.93633277 0.02356719 0.4227285  0.36964614  0.94248007
#> b[(Intercept) herd:8]   0.54044563 0.02505708 0.4714153 -0.03495114  0.53504744
#> b[(Intercept) herd:9]  -0.29855559 0.02752184 0.5994836 -1.05611929 -0.29690968
#> b[(Intercept) herd:10] -0.63742302 0.02256440 0.4515056 -1.23389287 -0.61740557
#> b[(Intercept) herd:11] -0.14289200 0.01883113 0.3808413 -0.63791759 -0.13608123
#> b[(Intercept) herd:12] -0.06555572 0.02220200 0.5301830 -0.76243882 -0.07447618
#> b[(Intercept) herd:13] -0.81596071 0.02239910 0.4702956 -1.46290565 -0.75003938
#> b[(Intercept) herd:14]  1.02070607 0.02133722 0.4567096  0.45202467  0.99353243
#> b[(Intercept) herd:15] -0.60039245 0.01886562 0.4654002 -1.17121616 -0.55848015
#>                                90% n_eff      Rhat
#> b[(Intercept) herd:1]   1.21997676   383 0.9966435
#> b[(Intercept) herd:2]   0.18312166   551 1.0001733
#> b[(Intercept) herd:3]   0.83592239   410 1.0015732
#> b[(Intercept) herd:4]   0.61272869   394 0.9972941
#> b[(Intercept) herd:5]   0.29051561   539 1.0000865
#> b[(Intercept) herd:6]   0.02560707   544 1.0002528
#> b[(Intercept) herd:7]   1.47831133   322 1.0006741
#> b[(Intercept) herd:8]   1.10466613   354 1.0026239
#> b[(Intercept) herd:9]   0.49438671   474 0.9984309
#> b[(Intercept) herd:10] -0.08642849   400 0.9993428
#> b[(Intercept) herd:11]  0.32962385   409 0.9980906
#> b[(Intercept) herd:12]  0.62454688   570 0.9974849
#> b[(Intercept) herd:13] -0.26039584   441 0.9986466
#> b[(Intercept) herd:14]  1.63921436   458 1.0000451
#> b[(Intercept) herd:15] -0.05283553   609 0.9973613