Summary method for stanreg objects

Summaries of parameter estimates and MCMC convergence diagnostics (Monte Carlo error, effective sample size, Rhat).

# S3 method for stanreg
summary(object, pars = NULL, regex_pars = NULL,
  probs = NULL, ..., digits = 1)

# S3 method for summary.stanreg
print(x, digits = max(1, attr(x, "print.digits")),
  ...)

# S3 method for summary.stanreg
as.data.frame(x, ...)

Arguments

object
A fitted model object returned by one of the rstanarm modeling functions. See stanreg-objects.

pars
An optional character vector specifying a subset of parameters to display. Parameters can be specified by name or several shortcuts can be used. Using pars="beta" will restrict the displayed parameters to only the regression coefficients (without the intercept). "alpha" can also be used as a shortcut for "(Intercept)". If the model has varying intercepts and/or slopes they can be selected using pars = "varying". If pars is NULL all parameters are selected. See Examples.

regex_pars
An optional character vector of regular expressions to use for parameter selection. regex_pars can be used in place of pars or in addition to pars. Currently, all functions that accept a regex_pars argument ignore it for models fit using optimization.

probs
For models fit using MCMC or one of the variational algorithms, an optional numeric vector of probabilities passed to quantile.

...
Currently ignored.

digits
Number of digits to use for formatting numbers when printing. When calling summary, the value of digits is stored as the "print.digits" attribute of the returned object.

x
An object of class "summary.stanreg".

Value

The summary method returns an object of class "summary.stanreg", which is a matrix of summary statistics and diagnostics, with attributes storing information for use by the print method. The print method for summary.stanreg objects is called for its side effect and just returns its input. The as.data.frame method for summary.stanreg objects converts the matrix to a data.frame, preserving row and column names but dropping the print-related attributes.

See also

prior_summary to extract or print a summary of the priors used for a particular model.

Examples

if (!exists("example_model")) example(example_model)

#> Warning: no help found for ‘example_model’

summary(example_model, probs = c(0.1, 0.9))

#> Error in summary(example_model, probs = c(0.1, 0.9)): object 'example_model' not found

# These produce the same output for this example, 
# but the second method can be used for any model
summary(example_model, pars = c("(Intercept)", "size",
                                paste0("period", 2:4)))

#> Error in summary(example_model, pars = c("(Intercept)", "size", paste0("period",     2:4))): object 'example_model' not found

summary(example_model, pars = c("alpha", "beta"))

#> Error in summary(example_model, pars = c("alpha", "beta")): object 'example_model' not found

# Only show parameters varying by group
summary(example_model, pars = "varying")

#> Error in summary(example_model, pars = "varying"): object 'example_model' not found

as.data.frame(summary(example_model, pars = "varying"))

#> Error in summary(example_model, pars = "varying"): object 'example_model' not found